Calculate counter-flow and net migration flow

Calculate counter-flow and net migration flow

counter(
  m,
  label = "flow",
  orig_col = "orig",
  dest_col = "dest",
  flow_col = "flow"
)

Arguments

m
A matrix or data frame of origin-destination flows. For matrix the first and second dimensions correspond to origin and destination respectively. For a data frame ensure the correct column names are passed to orig_col, dest_col and flow_col.

label
Character string for the prefix of the calculated columns. Can take values flow or stream

orig_col
Character string of the origin column name (when m is a data frame rather than a matrix)

dest_col
Character string of the destination column name (when m is a data frame rather than a matrix)

flow_col
Character string of the flow column name (when m is a data frame rather than a matrix)

Value

A tibble with columns for orig, destination, corridor, flow, counter-flow and net flow in each bilateral pair.

Examples

# matrix
r <- LETTERS[1:4]
m <- matrix(data = c(0, 100, 30, 70, 50, 0, 45, 5, 60, 35, 0, 40, 20, 25, 20, 0),
            nrow = 4, ncol = 4, dimnames = list(orig = r, dest = r), byrow = TRUE)
counter(m)

#> # A tibble: 12 x 7
#>    orig  dest  corridor pair   flow counter_flow net_flow
#>    <chr> <chr> <chr>    <chr> <dbl>        <dbl>    <dbl>
#>  1 B     A     B -> A   A - B    50          100      -50
#>  2 C     A     C -> A   A - C    60           30       30
#>  3 D     A     D -> A   A - D    20           70      -50
#>  4 A     B     A -> B   A - B   100           50       50
#>  5 C     B     C -> B   B - C    35           45      -10
#>  6 D     B     D -> B   B - D    25            5       20
#>  7 A     C     A -> C   A - C    30           60      -30
#>  8 B     C     B -> C   B - C    45           35       10
#>  9 D     C     D -> C   C - D    20           40      -20
#> 10 A     D     A -> D   A - D    70           20       50
#> 11 B     D     B -> D   B - D     5           25      -20
#> 12 C     D     C -> D   C - D    40           20       20

# data frame
library(dplyr)

#> 
#> Attaching package: 'dplyr'

#> The following objects are masked from 'package:stats':
#> 
#>     filter, lag

#> The following objects are masked from 'package:base':
#> 
#>     intersect, setdiff, setequal, union

library(tidyr)
d <- expand_grid(orig = r, dest = r, sex = c("female", "male")) %>%
  mutate(flow = sample(x = 1:100, size = 32))
d

#> # A tibble: 32 x 4
#>    orig  dest  sex     flow
#>    <chr> <chr> <chr>  <int>
#>  1 A     A     female    45
#>  2 A     A     male      23
#>  3 A     B     female    76
#>  4 A     B     male      63
#>  5 A     C     female    47
#>  6 A     C     male      31
#>  7 A     D     female    68
#>  8 A     D     male      73
#>  9 B     A     female    69
#> 10 B     A     male       5
#> # ... with 22 more rows

# use group_by to distinguish od tables
d %>%
  group_by(sex) %>%
  counter()

#> # A tibble: 24 x 8
#> # Groups:   sex [2]
#>    orig  dest  corridor pair  sex     flow counter_flow net_flow
#>    <chr> <chr> <chr>    <chr> <chr>  <int>        <int>    <int>
#>  1 A     B     A -> B   A - B female    76           69        7
#>  2 A     B     A -> B   A - B male      63            5       58
#>  3 A     C     A -> C   A - C female    47           43        4
#>  4 A     C     A -> C   A - C male      31           86      -55
#>  5 A     D     A -> D   A - D female    68           97      -29
#>  6 A     D     A -> D   A - D male      73            4       69
#>  7 B     A     B -> A   A - B female    69           76       -7
#>  8 B     A     B -> A   A - B male       5           63      -58
#>  9 B     C     B -> C   B - C female    77           84       -7
#> 10 B     C     B -> C   B - C male       2           91      -89
#> # ... with 14 more rows