Dictionary to look up region geographies based on countries used in UN DESA International Migrant Stock.

Intended for use as a custom dictionary with the countrycode package, where the existing UN region and area codes do not match those used by UN DESA in the WPP, see https://github.com/vincentarelbundock/countrycode/issues/253

dict_ims

Format

Data frame with 237 rows and 13 columns. One of first three columns intended as input for origin in countrycode.

name

Country name

iso3n

ISO 3 letter code

iso3c

ISO numeric code

Remaining columns intended as input for destination in countrycode.

region

Geographic region of country (6)

region_sub

Geographic sub region of country (22). Filled using region if none given in original data

region_sdg

SDG region of country (8)

region_sdg_sub

Sub SDG region of country (9). Filled using region_sdg if none given in original data

un_develop

UN development group of country (3)

wb_income

World Bank income group of country (3)

wb_income_detail

Detailled World Bank income group of country (4)

lldc

Indicator variable for Land-Locked Developing Countries (32)

sids

Indicator variable for Small Island Developing States (58)

region_as2014

Region grouping used for global chord diagram plots by Abel and Sander (2014)

region_sba2014

Region grouping used for global chord diagram plots by Sander, Abel and Bauer (2014)

region_a2018

Region grouping used for global chord diagram plots by Abel (2018)

region_ac2021

Region grouping used for global chord diagram plots by Abel and Cohen (2021)

Source

The aggregates_correspondence_table_2020_1.xlsx file of United Nations Department of Economic and Social Affairs, Population Division (2020). International Migrant Stock 2020.

Examples

# \dontrun{
library(tidyverse)

#> -- Attaching packages --------------------------------------- tidyverse 1.3.1 --

#> v ggplot2 3.3.5     v purrr   0.3.4
#> v tibble  3.1.5     v dplyr   1.0.7
#> v tidyr   1.1.4     v stringr 1.4.0
#> v readr   2.0.2     v forcats 0.5.1

#> -- Conflicts ------------------------------------------ tidyverse_conflicts() --
#> x dplyr::filter() masks stats::filter()
#> x dplyr::lag()    masks stats::lag()

library(countrycode)
# download Abel and Cohen (2019) estimates
f <- read_csv("https://ndownloader.figshare.com/files/26239945")

#> Error in (function (con, what, n = 1L, size = NA_integer_, signed = TRUE,     endian = .Platform$endian) {    if (!endian %in% c("big", "little", "swap"))         stop("invalid 'endian' argument")    if (is.character(con)) {        con <- file(con, "rb")        on.exit(close(con))    }    swap <- endian != .Platform$endian    if (!is.character(what) || is.na(what) || length(what) !=         1L || !any(what == c("numeric", "double", "integer",         "int", "logical", "complex", "character", "raw")))         what <- typeof(what)    .Internal(readBin(con, what, n, size, signed, swap))})(structure(4L, class = c("curl", "connection"), conn_id = <pointer: 0x000000000000058f>),     raw(0), 131071L): Send failure: Connection was reset

# use dictionary to get region to region flows
d <- f %>%
  mutate(
    orig = countrycode(
      sourcevar = orig, custom_dict = dict_ims,
      origin = "iso3c", destination = "region"),
    dest = countrycode(
      sourcevar = dest, custom_dict = dict_ims,
      origin = "iso3c", destination = "region")
  ) %>%
  group_by(year0, orig, dest) %>%
  summarise_all(sum)

#> Error in mutate(., orig = countrycode(sourcevar = orig, custom_dict = dict_ims,     origin = "iso3c", destination = "region"), dest = countrycode(sourcevar = dest,     custom_dict = dict_ims, origin = "iso3c", destination = "region")): object 'f' not found

d

#> Error in eval(expr, envir, enclos): object 'd' not found

# }