Estimation of bilateral migrant flows from bilateral migrant stocks using demographic accounting approaches

Estimates migrant transitions flows between two sequential migrant stock tables. Replaces old ffs.

ffs_demo(
  m1 = NULL,
  m2 = NULL,
  b_por = NULL,
  d_por = NULL,
  m = NULL,
  stayer_assumption = TRUE,
  match_global = "before-demo-adjust",
  match_pob_tot_method = "rescale",
  birth_method = "native",
  birth_non_negative = TRUE,
  death_method = "proportion",
  verbose = FALSE,
  ...
)

Arguments

m1
Matrix of migrant stock totals at time t. Rows in the matrix correspond to place of birth and columns to place of residence at time t

m2
Matrix of migrant stock totals at time t+1. Rows in the matrix correspond to place of birth and columns to place of residence at time t+1.

b_por
Vector of the number of births between time t and t+1 in each region.

d_por
Vector of the number of deaths between time t and t+1 in each region.

m
Matrix of auxiliary data. By default set to 1 for all origin-destination combinations.

stayer_assumption
Logical value to indicate whether to use ipf3 or ipf3_qi to estimate flows. By default uses ipf3_qi, i.e. is set to TRUE. The ipf function is useful for replicating method of Azose and Raftery.

match_global
Character string used to indicate whether to balance the change in stocks totals with the changes in births and deaths. Only applied when match_pob_tot_method is either rescale or rescale-adjust-zero-fb. By default uses after-demo-adjust rather than before-demo-adjust which I think minimises risk of negative values.

match_pob_tot_method
Character string passed to method argument in match_pob_tot to ensure place of birth margins in stock tables match.

birth_method
Character string passed to method argument in birth_mat.

birth_non_negative
Logical value passed to non_negative argument in birth_mat.

death_method
Character string passed to method argument in death_mat.

verbose
Logical value to indicate the print the parameter estimates at each iteration of the various IPF routines. By default FALSE.

...
Additional arguments passes to ipf3_qi or ipf3.

Value

Estimates migrant transitions flows between two sequential migrant stock tables using various methods. See the example section for possible variations on estimation methods.

Returns a list object with:

mu

Array of indirect estimates of origin-destination matrices by place of birth.

it

Iteration count.

tol

Tolerance level at final iteration.

y

Array of indirect estimates of origin-destination matrices by place of birth with additional rows and columns for births, deaths and moves to other regions.

...

Slots to record which estimation method was used (as set by arguments above)

od_flow

Matrix of estimated origin-destination flows

References

Abel and Cohen (2019) Bilateral international migration flow estimates for 200 countries Scientific Data 6 (1), 1-13

Azose & Raftery (2019) Estimation of emigration, return migration, and transit migration between all pairs of countries Proceedings of the National Academy of Sciences 116 (1) 116-122

Abel, G. J. (2018). Estimates of Global Bilateral Migration Flows by Gender between 1960 and 2015. International Migration Review 52 (3), 809–852.

Abel, G. J. and Sander, N. (2014). Quantifying Global International Migration Flows. Science, 343 (6178) 1520-1522

Abel, G. J. (2013). Estimating Global Migration Flow Tables Using Place of Birth. Demographic Research 28, (18) 505-546

See also

Author

Guy J. Abel

Examples

##
## without births and deaths over period
##
# data as in papers
s1 <- matrix(data = c(1000, 100, 10, 0, 55, 555, 50, 5, 80, 40, 800, 40, 20, 25, 20, 200),
             nrow = 4, ncol = 4, byrow = TRUE)
s2 <- matrix(data = c(950, 100, 60, 0, 80, 505, 75, 5, 90, 30, 800, 40, 40, 45, 0, 180),
             nrow = 4, ncol = 4, byrow = TRUE)
b <- d <- rep(0, 4)
r <- LETTERS[1:4]
dimnames(s1) <- dimnames(s2) <- list(pob = r, por = r)
names(b) <- names(d) <- r
s1

#>    por
#> pob    A   B   C   D
#>   A 1000 100  10   0
#>   B   55 555  50   5
#>   C   80  40 800  40
#>   D   20  25  20 200

s2

#>    por
#> pob   A   B   C   D
#>   A 950 100  60   0
#>   B  80 505  75   5
#>   C  90  30 800  40
#>   D  40  45   0 180

b

#> A B C D 
#> 0 0 0 0 

d

#> A B C D 
#> 0 0 0 0 

# demographic research and science paper example
e0 <- ffs_demo(m1 = s1, m2 = s2, b_por = b, d_por = d)
e0$od_flow

#>      dest
#> orig   A  B  C D Sum
#>   A    0  0 50 0  50
#>   B   35  0 25 0  60
#>   C   10 10  0 0  20
#>   D   10 10  0 0  20
#>   Sum 55 20 75 0 150

# international migration review paper example
s1[,] <- c(100, 20, 10, 20, 10, 55, 40, 25, 10, 25, 140, 20, 0, 10, 65, 200)
s2[,] <- c(70, 25, 10, 40, 30, 60, 55, 45, 10, 10, 140, 0, 10, 15, 50, 180)
e1 <- ffs_demo(m1 = s1, m2 = s2, b_por = b, d_por = d)

#> Error in ipf3_qi(row_tot = t(m1_d), col_tot = m2_d, m = m, verbose = verbose,     ...): row and column totals are not equal for one or more sub-tables, ensure colSums(row_tot)==rowSums(col_tot)

e1$od_flow

#> Error in eval(expr, envir, enclos): object 'e1' not found

# international migration review supp. material example
# distance matrix
dd <- matrix(data = c(0, 5, 50, 500, 5, 0, 45, 495, 50, 45, 0, 450, 500, 495, 450, 0), 
             nrow = 4, ncol = 4, byrow = TRUE)
dimnames(dd) <- list(orig = r, dest = r)
dd

#>     dest
#> orig   A   B   C   D
#>    A   0   5  50 500
#>    B   5   0  45 495
#>    C  50  45   0 450
#>    D 500 495 450   0

e3 <- ffs_demo(m1 = s1, m2 = s2, b_por = b, d_por = d, m = dd)
e3$od_flow

#>      dest
#> orig         A        B C  D Sum
#>   A    0.00000 20.00000 0 10  30
#>   B    0.00000  0.00000 0  0   0
#>   C   15.23823 14.76177 0  5  35
#>   D    9.76177 25.23823 0  0  35
#>   Sum 25.00000 60.00000 0 15 100

##
## with births and deaths over period
##
# demographic research paper example (with births and deaths)
s1[,] <- c(1000, 55, 80, 20, 100, 555, 40, 25, 10, 50, 800, 20, 0, 5, 40, 200)
s2[,] <- c(1060, 45, 70, 30, 60, 540, 75, 30, 10, 40, 770, 20, 10, 0, 70, 230)
b[] <- c(80, 20, 40, 60)
d[] <- c(70, 30, 50, 10)
e4 <- ffs_demo(m1 = s1, m2 = s2, b_por = b, d_por = d, match_pob_tot_method = "open-dr")
# makes more sense to use this method
e5 <- ffs_demo(m1 = s1, m2 = s2, b_por = b, d_por = d, match_pob_tot_method = "open")
e5$od_flow

#>      dest
#> orig         A         B         C         D       Sum
#>   A    0.00000  8.897732 0.0000000  3.852711  12.75044
#>   B   26.81098  0.000000 0.4296748 10.000000  37.24065
#>   C    0.00000 35.677466 0.0000000 26.145924  61.82339
#>   D   11.62571 11.251175 1.4120903  0.000000  24.28898
#>   Sum 38.43669 55.826373 1.8417650 39.998635 136.10346

# science paper  supp. material example
b[] <- c(80, 20, 60, 60)
e6 <- ffs_demo(m1 = s1, m2 = s2, b_por = b, d_por = d)
e6$od_flow

#>      dest
#> orig             A         B        C         D        Sum
#>   A    0.000000000  3.327696 0.000000  3.240618   6.568314
#>   B   34.107236105  0.000000 0.583247  9.811330  44.501813
#>   C    0.002843014 27.155724 0.000000 25.713389  52.871956
#>   D   12.434176193  9.076551 2.259055  0.000000  23.769782
#>   Sum 46.544255312 39.559971 2.842302 38.765337 127.711866

# international migration review supp. material example (with births and deaths)
s1[,] <- c(100, 20, 10, 20, 10, 55, 40, 25, 10, 25, 140, 20, 0, 10, 65, 200)
s2[,] <- c(75, 20, 30, 30, 25, 45, 40, 30, 5, 30, 150, 20, 0, 15, 60, 230)
b[] <- c(10, 50, 25, 60)
d[] <- c(30, 10, 40, 10)
e7 <- ffs_demo(m1 = s1, m2 = s2, b_por = b, d_por = d)
e7$od_flow

#>      dest
#> orig         A        B         C        D       Sum
#>   A    0.00000 7.912363  0.000000 0.000000  7.912363
#>   B   15.62572 0.000000 17.798757 5.691586 39.116065
#>   C    0.00000 1.209485  0.000000 0.000000  1.209485
#>   D   17.84553 0.000000  7.892298 0.000000 25.737830
#>   Sum 33.47125 9.121848 25.691055 5.691586 73.975743