Summary indices of migration distance

Summary indices of migration distance

index_distance(
  m = NULL,
  d = NULL,
  orig_col = "orig",
  dest_col = "dest",
  flow_col = "flow",
  dist_col = "dist",
  long = TRUE
)

Arguments

m
A matrix or data frame of origin-destination flows. For matrix the first and second dimensions correspond to origin and destination respectively. For a data frame ensure the correct column names are passed to orig_col, dest_col and flow_col.

d
A matrix or data frame of origin-destination distances. For matrix the first and second dimensions correspond to origin and destination respectively. For a data frame ensure the correct column names are passed to orig_col, dest_col and dist_col. Region names should match those in m.

orig_col
Character string of the origin column name (when m is a data frame rather than a matrix)

dest_col
Character string of the destination column name (when m is a data frame rather than a matrix)

flow_col
Character string of the flow column name (when m is a data frame rather than a matrix)

dist_col
Character string of the distance column name (when dist is a data frame rather than a matrix)

long
Logical to return a long data frame with index values all in one column

Source

Bell, M., Blake, M., Boyle, P., Duke-Williams, O., Rees, P. H., Stillwell, J., & Hugo, G. J. (2002). Cross-national comparison of internal migration: issues and measures. Journal of the Royal Statistical Society: Series A (Statistics in Society), 165(3), 435–464. https://doi.org/10.1111/1467-985X.00247

Value

A tibble with 3 summary measures where

mean

Mean migration distance from Bell et. al. (2002) - not discussed in text but given in Table 6

median

Mean migration distance from Bell et. al. (2002)

decay

Distance decay parameter obtained from a Poisson regression model (flow ~ orig + dest + log(dist))

Examples

# single year
index_distance(
  m = subset(korea_reg, year == 2020), 
  d = korea_dist
)

#> # A tibble: 3 x 2
#>   measure   value
#>   <chr>     <dbl>
#> 1 mean    105.   
#> 2 median   68.7  
#> 3 decay    -0.852

library(dplyr)
library(tidyr)
library(purrr)
# multiple years
korea_reg %>%
  nest(m = c(orig, dest, flow)) %>%
  mutate(d = list(korea_dist)) %>%
  mutate(i = map2(.x = m, .y = d, 
                  .f = ~index_distance(m = .x, d = .y, long = FALSE))) %>%
  select(-m, -d) %>%
  unnest(i)

#> # A tibble: 9 x 4
#>    year  mean median  decay
#>   <int> <dbl>  <dbl>  <dbl>
#> 1  2012  108.   76   -0.784
#> 2  2013  107.   76   -0.795
#> 3  2014  108.   76   -0.816
#> 4  2015  108.   76   -0.828
#> 5  2016  107.   76   -0.820
#> 6  2017  108.   75.8 -0.839
#> 7  2018  107.   74.8 -0.839
#> 8  2019  107.   75.8 -0.836
#> 9  2020  105.   68.7 -0.852