Lifetime migration totals for states and zones in the Indian 1901 to 1931

Lifetime migration (stock) totals from India

Usage

indian_sub

Format

Data frame with 164 rows and 7 columns:

zone

Zone of state. In some cases the state and zone are the same entity

state

Indian state

sex

Migrant sex

in_migrants

In-migrant total based on birthplace

out_migrants

Out-migrant total based on birthplace

net_migrants

Net migrant total based on birthplace

Source

Zachariah, K. C. (1964). A Historical Study of Internal Migration in the Indian Sub-Continent 1901-1931. (Vol. 19). Asia Publishing House.

Scraped from https://archive.org/details/in.ernet.dli.2015.130424/page/n73/mode/2up