Multiplicative component description of origin-destination migration flow tables

Multiplicative component descriptions of *n*-dimension flow tables based on total reference coding system.

multi_comp(m)

Arguments

m
matrix or array of migration flows

Value

matrix or array of multiplicative components of `m`. When output is an array the total for each table of origin-destination flows is used.

Examples

r <- LETTERS[1:4]
m0 <- matrix(data = c(0, 100, 30, 70, 50, 0, 45, 5, 60, 35, 0, 40, 20, 25, 20, 0), 
             nrow = 4, ncol = 4, byrow = TRUE, dimnames = list(orig = r, dest = r))
addmargins(m0)

#>      dest
#> orig    A   B  C   D Sum
#>   A     0 100 30  70 200
#>   B    50   0 45   5 100
#>   C    60  35  0  40 135
#>   D    20  25 20   0  65
#>   Sum 130 160 95 115 500

multi_comp(m = m0)

#>      dest
#> orig            A           B           C           D         Sum
#>   A     0.0000000   1.5625000   0.7894737   1.5217391   0.4000000
#>   B     1.9230769   0.0000000   2.3684211   0.2173913   0.2000000
#>   C     1.7094017   0.8101852   0.0000000   1.2882448   0.2700000
#>   D     1.1834320   1.2019231   1.6194332   0.0000000   0.1300000
#>   Sum   0.2600000   0.3200000   0.1900000   0.2300000 500.0000000

# data frame
library(dplyr)
italy_area %>%
  filter(year == 2000) %>%
  multi_comp()

#> , , age_grp = 0-4
#> 
#>             dest
#> orig               Center      Islands   North-East   North-West        South
#>   Center     0.000000e+00 1.401123e+00 8.588357e-01 9.089951e-01 2.369956e+00
#>   Islands    9.698334e-01 0.000000e+00 1.181023e+00 1.512619e+00 6.812558e-01
#>   North-East 1.053459e+00 1.915786e+00 0.000000e+00 1.178693e+00 2.501280e+00
#>   North-West 8.772374e-01 2.490386e+00 8.871662e-01 0.000000e+00 2.023312e+00
#>   South      1.408864e+00 5.307788e-01 1.183818e+00 1.101941e+00 0.000000e+00
#>   Sum        1.637855e-02 6.862844e-03 1.687236e-02 1.757883e-02 1.430240e-02
#>             dest
#> orig                  Sum
#>   Center     1.032671e-02
#>   Islands    1.241007e-02
#>   North-East 1.018613e-02
#>   North-West 1.418706e-02
#>   South      2.488502e-02
#>   Sum        7.199498e-02
#> 
#> , , age_grp = 5-9
#> 
#>             dest
#> orig               Center      Islands   North-East   North-West        South
#>   Center     0.000000e+00 1.588938e+00 7.792117e-01 7.616549e-01 2.242826e+00
#>   Islands    1.165630e+00 0.000000e+00 1.392577e+00 1.706670e+00 5.617293e-01
#>   North-East 8.400794e-01 1.932307e+00 0.000000e+00 9.360518e-01 2.084792e+00
#>   North-West 8.767757e-01 2.714381e+00 8.444276e-01 0.000000e+00 1.962796e+00
#>   South      1.386890e+00 5.069939e-01 1.283050e+00 1.150545e+00 0.000000e+00
#>   Sum        1.137199e-02 5.053418e-03 1.237763e-02 1.220101e-02 9.086780e-03
#>             dest
#> orig                  Sum
#>   Center     6.707853e-03
#>   Islands    9.717556e-03
#>   North-East 5.950922e-03
#>   North-West 9.847316e-03
#>   South      1.786718e-02
#>   Sum        5.009083e-02
#> 
#> , , age_grp = 10-14
#> 
#>             dest
#> orig               Center      Islands   North-East   North-West        South
#>   Center     0.000000e+00 1.570042e+00 7.377172e-01 6.665338e-01 1.978112e+00
#>   Islands    1.332680e+00 0.000000e+00 1.572373e+00 1.791277e+00 4.633897e-01
#>   North-East 8.613732e-01 1.833810e+00 0.000000e+00 8.404388e-01 1.805252e+00
#>   North-West 7.926695e-01 2.693539e+00 8.262637e-01 0.000000e+00 1.958533e+00
#>   South      1.424488e+00 4.108164e-01 1.331595e+00 1.226285e+00 0.000000e+00
#>   Sum        8.708315e-03 3.597226e-03 9.576984e-03 9.241771e-03 6.199628e-03
#>             dest
#> orig                  Sum
#>   Center     4.548797e-03
#>   Islands    7.799997e-03
#>   North-East 4.083825e-03
#>   North-West 7.136781e-03
#>   South      1.375452e-02
#>   Sum        3.732392e-02
#> 
#> , , age_grp = 15-19
#> 
#>             dest
#> orig               Center      Islands   North-East   North-West        South
#>   Center     0.000000e+00 1.357925e+00 7.321129e-01 6.675856e-01 1.673050e+00
#>   Islands    1.260990e+00 0.000000e+00 1.617163e+00 2.108876e+00 4.173974e-01
#>   North-East 6.770229e-01 1.768594e+00 0.000000e+00 8.473352e-01 1.696776e+00
#>   North-West 6.289080e-01 2.606491e+00 8.176568e-01 0.000000e+00 1.802664e+00
#>   South      1.449203e+00 3.469145e-01 1.449472e+00 1.340216e+00 0.000000e+00
#>   Sum        9.097594e-03 3.716172e-03 1.112689e-02 1.129269e-02 6.203232e-03
#>             dest
#> orig                  Sum
#>   Center     4.620886e-03
#>   Islands    9.144451e-03
#>   North-East 4.199167e-03
#>   North-West 7.306190e-03
#>   South      1.616589e-02
#>   Sum        4.143658e-02
#> 
#> , , age_grp = 20-24
#> 
#>             dest
#> orig               Center      Islands   North-East   North-West        South
#>   Center     0.000000e+00 1.043853e+00 8.515426e-01 7.590432e-01 1.551730e+00
#>   Islands    9.829151e-01 0.000000e+00 1.490343e+00 1.948493e+00 4.361107e-01
#>   North-East 5.925902e-01 1.530391e+00 0.000000e+00 8.520667e-01 1.807921e+00
#>   North-West 5.327866e-01 1.879603e+00 7.255385e-01 0.000000e+00 1.449256e+00
#>   South      1.419440e+00 4.250418e-01 1.788174e+00 1.623689e+00 0.000000e+00
#>   Sum        2.471561e-02 9.101198e-03 3.642281e-02 3.670756e-02 1.705979e-02
#>             dest
#> orig                  Sum
#>   Center     1.396719e-02
#>   Islands    2.456422e-02
#>   North-East 1.226229e-02
#>   North-West 1.771580e-02
#>   South      5.549748e-02
#>   Sum        1.240070e-01
#> 
#> , , age_grp = 25-29
#> 
#>             dest
#> orig               Center      Islands   North-East   North-West        South
#>   Center     0.000000e+00 1.092255e+00 9.918236e-01 9.388465e-01 2.092521e+00
#>   Islands    9.148755e-01 0.000000e+00 1.220538e+00 1.599428e+00 5.437596e-01
#>   North-East 9.099640e-01 1.419510e+00 0.000000e+00 1.161175e+00 1.829372e+00
#>   North-West 7.946255e-01 1.652364e+00 9.473912e-01 0.000000e+00 1.718638e+00
#>   South      1.472875e+00 4.815876e-01 1.457440e+00 1.373074e+00 0.000000e+00
#>   Sum        4.392364e-02 1.394556e-02 5.303926e-02 5.408094e-02 3.226330e-02
#>             dest
#> orig                  Sum
#>   Center     2.739010e-02
#>   Islands    3.389971e-02
#>   North-East 2.318012e-02
#>   North-West 3.388169e-02
#>   South      7.890108e-02
#>   Sum        1.972527e-01
#> 
#> , , age_grp = 30-34
#> 
#>             dest
#> orig               Center      Islands   North-East   North-West        South
#>   Center     0.000000e+00 1.210655e+00 1.087738e+00 1.159314e+00 2.136373e+00
#>   Islands    9.149807e-01 0.000000e+00 1.053335e+00 1.389918e+00 5.263060e-01
#>   North-East 1.142543e+00 1.383664e+00 0.000000e+00 1.362224e+00 1.836868e+00
#>   North-West 9.447906e-01 1.856680e+00 1.091411e+00 0.000000e+00 1.756123e+00
#>   South      1.543976e+00 4.450803e-01 1.205220e+00 1.243742e+00 0.000000e+00
#>   Sum        3.882697e-02 1.187301e-02 4.001644e-02 4.290719e-02 2.651783e-02
#>             dest
#> orig                  Sum
#>   Center     2.458945e-02
#>   Islands    2.485258e-02
#>   North-East 2.092014e-02
#>   North-West 3.056921e-02
#>   South      5.921005e-02
#>   Sum        1.601414e-01
#> 
#> , , age_grp = 35-39
#> 
#>             dest
#> orig               Center      Islands   North-East   North-West        South
#>   Center     0.000000e+00 1.438609e+00 1.175195e+00 1.245360e+00 2.125698e+00
#>   Islands    9.563470e-01 0.000000e+00 1.073211e+00 1.371731e+00 4.066445e-01
#>   North-East 1.278130e+00 1.395618e+00 0.000000e+00 1.483509e+00 1.718642e+00
#>   North-West 1.157817e+00 2.026330e+00 1.228607e+00 0.000000e+00 1.752761e+00
#>   South      1.464728e+00 4.236468e-01 1.089155e+00 1.085477e+00 0.000000e+00
#>   Sum        2.409565e-02 7.601753e-03 2.400554e-02 2.487421e-02 1.527920e-02
#>             dest
#> orig                  Sum
#>   Center     1.559639e-02
#>   Islands    1.471330e-02
#>   North-East 1.305526e-02
#>   North-West 2.019204e-02
#>   South      3.229934e-02
#>   Sum        9.585634e-02
#> 
#> , , age_grp = 40-44
#> 
#>             dest
#> orig               Center      Islands   North-East   North-West        South
#>   Center     0.000000e+00 1.546986e+00 1.283416e+00 1.266086e+00 2.199505e+00
#>   Islands    1.000570e+00 0.000000e+00 1.090087e+00 1.444766e+00 3.665541e-01
#>   North-East 1.321963e+00 1.562994e+00 0.000000e+00 1.416881e+00 1.625943e+00
#>   North-West 1.234426e+00 2.352920e+00 1.260763e+00 0.000000e+00 1.884671e+00
#>   South      1.317262e+00 3.536289e-01 1.043553e+00 1.000530e+00 0.000000e+00
#>   Sum        1.314537e-02 4.592050e-03 1.352384e-02 1.355628e-02 8.704710e-03
#>             dest
#> orig                  Sum
#>   Center     9.137242e-03
#>   Islands    8.456004e-03
#>   North-East 7.230496e-03
#>   North-West 1.208927e-02
#>   South      1.660924e-02
#>   Sum        5.352225e-02
#> 
#> , , age_grp = 45-49
#> 
#>             dest
#> orig               Center      Islands   North-East   North-West        South
#>   Center     0.000000e+00 1.638414e+00 1.129906e+00 1.204070e+00 2.331083e+00
#>   Islands    1.076053e+00 0.000000e+00 1.099888e+00 1.372331e+00 3.999248e-01
#>   North-East 1.406491e+00 1.701142e+00 0.000000e+00 1.501220e+00 1.607074e+00
#>   North-West 1.320154e+00 2.600060e+00 1.353659e+00 0.000000e+00 2.007255e+00
#>   South      1.285796e+00 4.075609e-01 9.188811e-01 9.118458e-01 0.000000e+00
#>   Sum        8.308222e-03 3.099814e-03 7.926152e-03 7.969406e-03 5.615710e-03
#>             dest
#> orig                  Sum
#>   Center     5.496763e-03
#>   Islands    5.237244e-03
#>   North-East 4.653325e-03
#>   North-West 7.998241e-03
#>   South      9.533730e-03
#>   Sum        3.291930e-02
#> 
#> , , age_grp = 50-54
#> 
#>             dest
#> orig               Center      Islands   North-East   North-West        South
#>   Center     0.000000e+00 1.886952e+00 1.064421e+00 1.110475e+00 2.578888e+00
#>   Islands    9.972860e-01 0.000000e+00 8.610614e-01 1.226001e+00 3.613748e-01
#>   North-East 1.448846e+00 1.709439e+00 0.000000e+00 1.504786e+00 1.541422e+00
#>   North-West 1.518892e+00 3.174323e+00 1.594591e+00 0.000000e+00 2.391048e+00
#>   South      1.266925e+00 3.658834e-01 7.375885e-01 8.307348e-01 0.000000e+00
#>   Sum        7.104341e-03 2.908779e-03 6.163584e-03 6.174397e-03 5.150737e-03
#>             dest
#> orig                  Sum
#>   Center     4.678556e-03
#>   Islands    3.809888e-03
#>   North-East 3.889185e-03
#>   North-West 7.900921e-03
#>   South      7.223288e-03
#>   Sum        2.750184e-02
#> 
#> , , age_grp = 55-59
#> 
#>             dest
#> orig               Center      Islands   North-East   North-West        South
#>   Center     0.000000e+00 2.262956e+00 1.084147e+00 1.029319e+00 2.894385e+00
#>   Islands    8.453738e-01 0.000000e+00 6.431451e-01 1.027474e+00 3.434912e-01
#>   North-East 1.448005e+00 1.640573e+00 0.000000e+00 1.454790e+00 1.391090e+00
#>   North-West 1.723596e+00 3.928920e+00 1.891787e+00 0.000000e+00 2.921378e+00
#>   South      1.160023e+00 3.984941e-01 5.440627e-01 7.223716e-01 0.000000e+00
#>   Sum        5.705820e-03 2.782624e-03 4.808316e-03 4.527170e-03 4.721810e-03
#>             dest
#> orig                  Sum
#>   Center     4.062198e-03
#>   Islands    2.577171e-03
#>   North-East 3.063770e-03
#>   North-West 7.727908e-03
#>   South      5.114693e-03
#>   Sum        2.254574e-02
#> 
#> , , age_grp = 60-64
#> 
#>             dest
#> orig               Center      Islands   North-East   North-West        South
#>   Center     0.000000e+00 2.270536e+00 1.066661e+00 1.084406e+00 3.282451e+00
#>   Islands    7.666707e-01 0.000000e+00 3.969951e-01 9.328310e-01 4.139080e-01
#>   North-East 1.330674e+00 1.472774e+00 0.000000e+00 1.578228e+00 1.499957e+00
#>   North-West 1.632759e+00 4.038468e+00 1.937749e+00 0.000000e+00 3.046622e+00
#>   South      1.245437e+00 3.951250e-01 4.437330e-01 7.338774e-01 0.000000e+00
#>   Sum        5.446301e-03 2.696117e-03 4.267651e-03 4.451477e-03 4.966911e-03
#>             dest
#> orig                  Sum
#>   Center     4.177540e-03
#>   Islands    2.155452e-03
#>   North-East 2.991681e-03
#>   North-West 7.572918e-03
#>   South      4.930867e-03
#>   Sum        2.182846e-02
#> 
#> , , age_grp = 65-69
#> 
#>             dest
#> orig               Center      Islands   North-East   North-West        South
#>   Center     0.000000e+00 2.383049e+00 1.158613e+00 1.030143e+00 3.451178e+00
#>   Islands    8.265242e-01 0.000000e+00 4.351265e-01 8.755056e-01 3.850425e-01
#>   North-East 1.222004e+00 1.237478e+00 0.000000e+00 1.628584e+00 1.439415e+00
#>   North-West 1.518239e+00 3.323839e+00 1.890634e+00 0.000000e+00 2.933009e+00
#>   South      1.339701e+00 4.788613e-01 4.192163e-01 8.738530e-01 0.000000e+00
#>   Sum        4.209980e-03 1.892328e-03 3.283640e-03 3.590017e-03 3.784657e-03
#>             dest
#> orig                  Sum
#>   Center     3.323289e-03
#>   Islands    1.665249e-03
#>   North-East 2.209519e-03
#>   North-West 5.417466e-03
#>   South      4.145100e-03
#>   Sum        1.676062e-02
#> 
#> , , age_grp = 70-74
#> 
#>             dest
#> orig               Center      Islands   North-East   North-West        South
#>   Center     0.000000e+00 2.408722e+00 9.994382e-01 1.353416e+00 3.269315e+00
#>   Islands    7.229609e-01 0.000000e+00 3.812449e-01 1.253460e+00 3.853316e-01
#>   North-East 1.301347e+00 1.200007e+00 0.000000e+00 1.765216e+00 1.112676e+00
#>   North-West 1.420825e+00 2.608192e+00 1.718798e+00 0.000000e+00 2.445270e+00
#>   South      1.451447e+00 4.324798e-01 4.602659e-01 1.064537e+00 0.000000e+00
#>   Sum        3.662106e-03 1.430961e-03 2.631237e-03 3.827910e-03 2.851108e-03
#>             dest
#> orig                  Sum
#>   Center     2.890757e-03
#>   Islands    1.578742e-03
#>   North-East 1.892328e-03
#>   North-West 4.051385e-03
#>   South      3.990109e-03
#>   Sum        1.440332e-02
#> 
#> , , age_grp = 75-79
#> 
#>             dest
#> orig               Center      Islands   North-East   North-West        South
#>   Center     0.000000e+00 1.925821e+00 1.173672e+00 1.310937e+00 2.957355e+00
#>   Islands    8.188638e-01 0.000000e+00 3.517830e-01 1.352001e+00 4.311971e-01
#>   North-East 1.394874e+00 8.398864e-01 0.000000e+00 2.114284e+00 9.289683e-01
#>   North-West 1.326848e+00 2.463358e+00 1.809854e+00 0.000000e+00 1.963189e+00
#>   South      1.450123e+00 4.365768e-01 4.882748e-01 1.172861e+00 0.000000e+00
#>   Sum        3.373751e-03 1.149815e-03 2.569962e-03 3.820701e-03 2.256376e-03
#>             dest
#> orig                  Sum
#>   Center     2.526709e-03
#>   Islands    1.549907e-03
#>   North-East 1.816635e-03
#>   North-West 3.460258e-03
#>   South      3.817097e-03
#>   Sum        1.317061e-02
#> 
#> , , age_grp = 80-84
#> 
#>             dest
#> orig               Center      Islands   North-East   North-West        South
#>   Center     0.000000e+00 1.845544e+00 1.070267e+00 1.502951e+00 2.636205e+00
#>   Islands    8.035932e-01 0.000000e+00 4.282419e-01 1.294605e+00 5.185149e-01
#>   North-East 1.465781e+00 6.312277e-01 0.000000e+00 2.117016e+00 9.859721e-01
#>   North-West 1.231642e+00 2.000778e+00 1.825466e+00 0.000000e+00 1.826414e+00
#>   South      1.570968e+00 4.083209e-01 4.929168e-01 1.258303e+00 0.000000e+00
#>   Sum        2.162661e-03 6.199628e-04 1.607578e-03 2.490665e-03 1.333641e-03
#>             dest
#> orig                  Sum
#>   Center     1.531885e-03
#>   Islands    9.876152e-04
#>   North-East 1.142606e-03
#>   North-West 2.025692e-03
#>   South      2.526709e-03
#>   Sum        8.214507e-03
#> 
#> , , age_grp = 85-89
#> 
#>             dest
#> orig               Center      Islands   North-East   North-West        South
#>   Center     0.000000e+00 1.545305e+00 1.509310e+00 1.605552e+00 2.574579e+00
#>   Islands    7.392006e-01 0.000000e+00 3.829700e-01 1.345374e+00 4.140804e-01
#>   North-East 1.766067e+00 1.253654e+00 0.000000e+00 2.809491e+00 9.133976e-01
#>   North-West 1.090487e+00 1.667269e+00 1.943631e+00 0.000000e+00 1.395139e+00
#>   South      1.410014e+00 3.007639e-01 4.152831e-01 1.239599e+00 0.000000e+00
#>   Sum        1.787800e-03 4.974120e-04 1.499445e-03 2.368114e-03 1.016451e-03
#>             dest
#> orig                  Sum
#>   Center     1.445378e-03
#>   Islands    8.254156e-04
#>   North-East 1.257948e-03
#>   North-West 1.618391e-03
#>   South      2.022088e-03
#>   Sum        7.169221e-03
#> 
#> , , age_grp = 90-94
#> 
#>             dest
#> orig               Center      Islands   North-East   North-West        South
#>   Center     0.000000e+00 1.319321e+00 1.210708e+00 1.906475e+00 2.276625e+00
#>   Islands    8.091032e-01 0.000000e+00 4.181377e-01 1.033085e+00 3.594374e-01
#>   North-East 1.469137e+00 1.082987e+00 0.000000e+00 2.835207e+00 6.598503e-01
#>   North-West 1.494117e+00 1.634620e+00 2.215983e+00 0.000000e+00 1.777900e+00
#>   South      1.452459e+00 2.501963e-01 3.869167e-01 1.141827e+00 0.000000e+00
#>   Sum        7.641402e-04 1.802217e-04 6.055451e-04 9.191309e-04 4.036967e-04
#>             dest
#> orig                  Sum
#>   Center     5.514785e-04
#>   Islands    2.991681e-04
#>   North-East 4.613677e-04
#>   North-West 7.785579e-04
#>   South      7.821624e-04
#>   Sum        2.872735e-03
#> 
#> , , age_grp = 95+
#> 
#>             dest
#> orig               Center      Islands   North-East   North-West        South
#>   Center     0.000000e+00 8.855967e-01 1.076062e+00 1.503858e+00 2.207384e+00
#>   Islands    8.472551e-01 0.000000e+00 5.212550e-01 8.347210e-01 5.227585e-01
#>   North-East 1.382559e+00 1.750083e+00 0.000000e+00 2.340346e+00 6.979439e-01
#>   North-West 1.707368e+00 2.593481e+00 2.148587e+00 0.000000e+00 2.262523e+00
#>   South      1.393766e+00 4.851738e-01 5.225292e-01 9.654939e-01 0.000000e+00
#>   Sum        2.667282e-04 8.650644e-05 2.162661e-04 2.595193e-04 1.585951e-04
#>             dest
#> orig                  Sum
#>   Center     1.621996e-04
#>   Islands    1.045286e-04
#>   North-East 1.477818e-04
#>   North-West 3.063770e-04
#>   South      2.667282e-04
#>   Sum        9.876152e-04
#> 
#> , , age_grp = Sum
#> 
#>             dest
#> orig               Center      Islands   North-East   North-West        South
#>   Center     0.000000e+00 1.721118e-02 3.742485e-02 3.922346e-02 5.387188e-02
#>   Islands    3.772402e-02 0.000000e+00 4.807595e-02 6.707853e-02 1.346977e-02
#>   North-East 3.028807e-02 1.590637e-02 0.000000e+00 4.112300e-02 3.727707e-02
#>   North-West 4.518880e-02 3.734195e-02 5.599490e-02 0.000000e+00 6.325783e-02
#>   South      1.198547e-01 1.322828e-02 1.110454e-01 1.154140e-01 0.000000e+00
#>   Sum        2.330556e-01 8.368777e-02 2.525411e-01 2.628390e-01 1.678766e-01
#>             dest
#> orig                  Sum
#>   Center     1.477314e-01
#>   Islands    1.663483e-01
#>   North-East 1.245945e-01
#>   North-West 2.017835e-01
#>   South      3.595424e-01
#>   Sum        2.774360e+05
#>