Summary of bilateral flows, counter-flow and net migration flow

Summary of bilateral flows, counter-flow and net migration flow

Usage

sum_bilat(
  m,
  label = "flow",
  orig_col = "orig",
  dest_col = "dest",
  flow_col = "flow"
)

Arguments

m

A matrix or data frame of origin-destination flows. For matrix the first and second dimensions correspond to origin and destination respectively. For a data frame ensure the correct column names are passed to orig_col, dest_col and flow_col.

label

Character string for the prefix of the calculated columns. Can take values flow or stream

orig_col

Character string of the origin column name (when m is a data frame rather than a matrix)

dest_col

Character string of the destination column name (when m is a data frame rather than a matrix)

flow_col

Character string of the flow column name (when m is a data frame rather than a matrix)

Value

A tibble with columns for orig, destination, corridor, flow, counter-flow and net flow in each bilateral pair.

Examples

# matrix
r <- LETTERS[1:4]
m <- matrix(data = c(0, 100, 30, 70, 50, 0, 45, 5, 60, 35, 0, 40, 20, 25, 20, 0),
            nrow = 4, ncol = 4, dimnames = list(orig = r, dest = r), byrow = TRUE)
m           
#>     dest
#> orig  A   B  C  D
#>    A  0 100 30 70
#>    B 50   0 45  5
#>    C 60  35  0 40
#>    D 20  25 20  0
sum_bilat(m)
#> Error in loadNamespace(x): there is no package called ‘dplyr’

# data frame
library(dplyr)
#> Error in library(dplyr): there is no package called ‘dplyr’
library(tidyr)
#> Error in library(tidyr): there is no package called ‘tidyr’
d <- expand_grid(orig = r, dest = r, sex = c("female", "male")) %>%
  mutate(flow = sample(x = 1:100, size = 32))
#> Error in mutate(., flow = sample(x = 1:100, size = 32)): could not find function "mutate"
d
#> Error in eval(expr, envir, enclos): object 'd' not found

# use group_by to distinguish od tables
d %>%
  group_by(sex) %>%
  sum_bilat()
#> Error in group_by(., sex): could not find function "group_by"