Extract a classic origin-destination migration flow matrix.

Extract a classic origin-destination migration flow matrix from a more detailed dis-aggregation of flows stored in an (array) object.

sum_od(y)

Arguments

y
Array of origin-destination matrices, where the first and second dimensions correspond to origin and destination respectively. Higher dimension(s) refer to additional migrant characteristic(s).

Value

Matrix from summing over the first and second dimension. Set diagonals to zero.

Returns a matrix object of origin-destination flows

Author

Guy J. Abel

Examples

dn <- LETTERS[1:2]
y <- cm3(row_tot = c(18, 20) * 2, col_tot = c(16, 22) * 2, 
         m = array(c(5, 1, 2, 7, 4, 2, 5, 9), dim = c(2, 2, 2), 
                   dimnames = list(orig = dn, dest = dn, type = c("ILL", "HEALTHY"))))

#> iteration: 0 
#> alpha parameters: 
#> beta parameters: 1 1 
#> 
#> iteration: 1 
#> alpha parameters: 2.25 2.105263 
#> beta parameters: 1.204557 0.8900719 
#> max difference: 1.25 
#> 
#> iteration: 2 
#> alpha parameters: 2.108776 2.240291 
#> beta parameters: 1.245143 0.8694605 
#> max difference: 0.1412239 
#> 
#> iteration: 3 
#> alpha parameters: 2.081826 2.2667 
#> beta parameters: 1.253107 0.8654605 
#> max difference: 0.02694999 
#> 
#> iteration: 4 
#> alpha parameters: 2.076582 2.271864 
#> beta parameters: 1.254664 0.8646796 
#> max difference: 0.005244326 
#> 
#> iteration: 5 
#> alpha parameters: 2.075558 2.272874 
#> beta parameters: 1.254969 0.864527 
#> max difference: 0.001024239 
#> 
#> iteration: 6 
#> alpha parameters: 2.075357 2.273071 
#> beta parameters: 1.255028 0.8644972 
#> max difference: 0.0002001792 
#> 
#> iteration: 7 
#> alpha parameters: 2.075318 2.273109 
#> beta parameters: 1.25504 0.8644914 
#> max difference: 3.912878e-05 
#> 
#> iteration: 8 
#> alpha parameters: 2.075311 2.273117 
#> beta parameters: 1.255042 0.8644902 
#> max difference: 7.648661e-06 
#> 
#> iteration: 9 
#> alpha parameters: 2.075309 2.273118 
#> beta parameters: 1.255043 0.86449 
#> max difference: 1.495122e-06 
#> 
#> iteration: 10 
#> alpha parameters: 2.075309 2.273119 
#> beta parameters: 1.255043 0.86449 
#> max difference: 2.922594e-07 
#> 

round(addmargins(y$n))

#> , , type = ILL
#> 
#>      dest
#> orig   A  B Sum
#>   A   13  4  17
#>   B    3 14  17
#>   Sum 16 17  33
#> 
#> , , type = HEALTHY
#> 
#>      dest
#> orig   A  B Sum
#>   A   10  9  19
#>   B    6 18  23
#>   Sum 16 27  43
#> 
#> , , type = Sum
#> 
#>      dest
#> orig   A  B Sum
#>   A   23 13  36
#>   B    9 31  40
#>   Sum 32 44  76
#> 

round(addmargins(sum_od(y$n))) 

#>      dest
#> orig  A  B Sum
#>   A   0 13  13
#>   B   9  0   9
#>   Sum 9 13  21