Unilateral summaries of in-, out-, turnover and net-migration totals from an origin-destination migration flow matrix or data frame.

Unilateral summaries of in-, out-, turnover and net-migration totals from an origin-destination migration flow matrix or data frame.

Alias for sum_region() for international data

Alias for sum_region() with more general naming

Alias for sum_unilat() with more explicit naming

Usage

sum_region(
  m,
  drop_diagonal = TRUE,
  orig = "orig",
  dest = "dest",
  flow = "flow",
  international = FALSE,
  include_net = TRUE,
  na_rm = TRUE
)

sum_country(
  m,
  drop_diagonal = TRUE,
  orig = "orig",
  dest = "dest",
  flow = "flow",
  include_net = TRUE,
  international = TRUE,
  na_rm = TRUE
)

sum_unilat(
  m,
  drop_diagonal = TRUE,
  orig = "orig",
  dest = "dest",
  flow = "flow",
  include_net = TRUE,
  international = TRUE,
  na_rm = TRUE
)

sum_unilateral(
  m,
  drop_diagonal = TRUE,
  orig = "orig",
  dest = "dest",
  flow = "flow",
  include_net = TRUE,
  international = TRUE,
  na_rm = TRUE
)

Arguments

m

A matrix or data frame of origin-destination flows. For matrix the first and second dimensions correspond to origin and destination respectively. For a data frame ensure the correct column names are passed to orig, dest and flow.

drop_diagonal

Logical to indicate dropping of diagonal terms, where the origin and destination are the same, in the calculation of totals. Default TRUE.

orig

Character string of the origin column name (when m is a data frame rather than a matrix)

dest

Character string of the destination column name (when m is a data frame rather than a matrix)

flow

Character string of the flow column name (when m is a data frame rather than a matrix)

international

Logical to indicate if flows are international.

include_net

Logical to indicate inclusion of a net migration total column for each region, in addition to the total in- and out-flows. Default TRUE.

na_rm

Logical to indicate if to remove NA values in m when calculating in and out migration flow totals. Default set to TRUE.

Value

A tibble with total in-, out- and turnover of flows for each region.

Examples

# matrix
r <- LETTERS[1:4]
m <- matrix(data = c(0, 100, 30, 70, 50, 0, 45, 5, 60, 35, 0, 40, 20, 25, 20, 0),
            nrow = 4, ncol = 4, dimnames = list(orig = r, dest = r), byrow = TRUE)
m
#>     dest
#> orig  A   B  C  D
#>    A  0 100 30 70
#>    B 50   0 45  5
#>    C 60  35  0 40
#>    D 20  25 20  0
sum_region(m)
#> # A tibble: 4 × 5
#>   region out_mig in_mig  turn   net
#>   <chr>    <dbl>  <dbl> <dbl> <dbl>
#> 1 A          200    130   330   -70
#> 2 B          100    160   260    60
#> 3 C          135     95   230   -40
#> 4 D           65    115   180    50

# \dontrun{
# data frame (tidy) format
library(tidyverse)

# download Abel and Cohen (2019) estimates
f <- read_csv("https://ndownloader.figshare.com/files/38016762", show_col_types = FALSE)
f
#> # A tibble: 307,833 × 9
#>    year0 orig  dest  sd_drop_neg sd_rev_neg mig_rate da_min_open da_min_closed
#>    <dbl> <chr> <chr>       <dbl>      <dbl>    <dbl>       <dbl>         <dbl>
#>  1  1990 BDI   BDI             0          0      0          0               0 
#>  2  1990 COM   BDI             0          0      0          0               0 
#>  3  1990 DJI   BDI             0          0      0          0               0 
#>  4  1990 ERI   BDI             0          0      0          0               0 
#>  5  1990 ETH   BDI             0          0      0          0               0 
#>  6  1990 KEN   BDI            30         30     75.7       51.3           207.
#>  7  1990 MDG   BDI             0          0      0          0.03            0 
#>  8  1990 MWI   BDI             0          0      0          0               0 
#>  9  1990 MUS   BDI             0          0      0          0.06            0 
#> 10  1990 MYT   BDI             0          0      0          0               0 
#> # ℹ 307,823 more rows
#> # ℹ 1 more variable: da_pb_closed <dbl>

# single period
f %>%
  filter(year0 == 1990) %>%
  sum_country(flow = "da_pb_closed")
#> # A tibble: 225 × 5
#>    country     emi      imm     turn      net
#>    <chr>     <dbl>    <dbl>    <dbl>    <dbl>
#>  1 ABW       2509.   10317.   12826.    7807.
#>  2 AFG     308122. 3722805. 4030927. 3414683.
#>  3 AGO      79047.  280649.  359696.  201603.
#>  4 AIA        417.    1278.    1694.     861.
#>  5 ALB     335191.   34440.  369632. -300751.
#>  6 AND       4921.   12465.   17386.    7544 
#>  7 ARE     236548.  551815.  788364.  315267.
#>  8 ARG     443259.  379700.  822958.  -63559.
#>  9 ARM     610241.  155818.  766059. -454423.
#> 10 ASM       5537.    3189.    8727.   -2348.
#> # ℹ 215 more rows

# all periods using group_by
f %>%
  group_by(year0) %>%
  sum_country(flow = "da_pb_closed")
#> # A tibble: 1,359 × 6
#> # Groups:   year0 [6]
#>    year0 country     emi      imm     turn      net
#>    <dbl> <chr>     <dbl>    <dbl>    <dbl>    <dbl>
#>  1  1990 ABW       2509.   10317.   12826.    7807.
#>  2  1990 AFG     308122. 3722805. 4030927. 3414683.
#>  3  1990 AGO      79047.  280649.  359696.  201603.
#>  4  1990 AIA        417.    1278.    1694.     861.
#>  5  1990 ALB     335191.   34440.  369632. -300751.
#>  6  1990 AND       4921.   12465.   17386.    7544 
#>  7  1990 ARE     236548.  551815.  788364.  315267.
#>  8  1990 ARG     443259.  379700.  822958.  -63559.
#>  9  1990 ARM     610241.  155818.  766059. -454423.
#> 10  1990 ASM       5537.    3189.    8727.   -2348.
#> # ℹ 1,349 more rows
# }