Calculate turnover and total in-, out- and net-migration totals from an origin-destination migration flow matrix or data frame.

Calculate turnover and total in-, out- and net-migration totals from an origin-destination migration flow matrix or data frame.

sum_turnover(
  m,
  drop_diagonal = TRUE,
  include_net = TRUE,
  orig_col = "orig",
  dest_col = "dest",
  flow_col = "flow",
  type = "internal",
  international = FALSE,
  name_tot = FALSE
)

Arguments

m
A matrix or data frame of origin-destination flows. For matrix the first and second dimensions correspond to origin and destination respectively. For a data frame ensure the correct column names are passed to orig_col, dest_col and flow_col.

drop_diagonal
Logical to indicate dropping of diagonal terms, where the origin and destination are the same, in the calculation of totals. Default TRUE.

include_net
Logical to indicate inclusion of a net migration total column for each region, in addition to the total in- and out-flows. Default TRUE.

orig_col
Character string of the origin column name (when m is a data frame rather than a matrix)

dest_col
Character string of the destination column name (when m is a data frame rather than a matrix)

flow_col
Character string of the flow column name (when m is a data frame rather than a matrix)

type
Character string to indicate if flows are internal or international to indicate if to use region, tot_in_mig, tot_out_mig or country, tot_imm and tot_emi in output.

international
Logical to indicate if flows are international.

name_tot
Logical to prefix column names with `"tot_"` to match older versions. Default `FALSE`.

Value

A tibble with total in-, out- and turnover of flows for each region.

Examples

# matrix
r <- LETTERS[1:4]
m <- matrix(data = c(0, 100, 30, 70, 50, 0, 45, 5, 60, 35, 0, 40, 20, 25, 20, 0),
            nrow = 4, ncol = 4, dimnames = list(orig = r, dest = r), byrow = TRUE)
m

#>     dest
#> orig  A   B  C  D
#>    A  0 100 30 70
#>    B 50   0 45  5
#>    C 60  35  0 40
#>    D 20  25 20  0

sum_turnover(m)

#> # A tibble: 4 x 5
#>   region in_mig out_mig  turn   net
#>   <chr>   <dbl>   <dbl> <dbl> <dbl>
#> 1 A         130     200   330   -70
#> 2 B         160     100   260    60
#> 3 C          95     135   230   -40
#> 4 D         115      65   180    50

  
# different labels
sum_turnover(m, international = TRUE)

#> # A tibble: 4 x 5
#>   country   imm   emi  turn   net
#>   <chr>   <dbl> <dbl> <dbl> <dbl>
#> 1 A         130   200   330   -70
#> 2 B         160   100   260    60
#> 3 C          95   135   230   -40
#> 4 D         115    65   180    50

# \dontrun{
# data frame (tidy) format
library(tidyverse)

# download Abel and Cohen (2019) estimates
f <- read_csv("https://ndownloader.figshare.com/files/26239945")

#> 
#> -- Column specification --------------------------------------------------------
#> cols(
#>   year0 = col_double(),
#>   orig = col_character(),
#>   dest = col_character(),
#>   sd_drop_neg = col_double(),
#>   sd_rev_neg = col_double(),
#>   mig_rate = col_double(),
#>   da_min_open = col_double(),
#>   da_min_closed = col_double(),
#>   da_pb_closed = col_double()
#> )

# turnover for single period
f %>% 
  filter(year0 == 1990) %>%
  sum_turnover(flow_col = "da_pb_closed", type = "international")

#> # A tibble: 197 x 5
#>    country     imm    emi    turn     net
#>    <chr>     <dbl>  <dbl>   <dbl>   <dbl>
#>  1 BDI       61630 381611  443241 -319981
#>  2 COM        9009  12011   21020   -3002
#>  3 DJI       10949  55945   66894  -44996
#>  4 ERI       14633 329383  344016 -314750
#>  5 ETH     1635513 177334 1812847 1458179
#>  6 KEN      306517  84833  391350  221684
#>  7 MDG        9706  19159   28865   -9453
#>  8 MWI      112416 974278 1086694 -861862
#>  9 MUS       16862  22475   39337   -5613
#> 10 MYT       13763   3021   16784   10742
#> # ... with 187 more rows

# turnover for all periods using group_by
f %>% 
  group_by(year0) %>%
  sum_turnover(flow_col = "da_pb_closed", type = "international")

#> Adding missing grouping variables: `year0`

#> # A tibble: 1,188 x 6
#> # Groups:   year0 [6]
#>    year0 country     imm    emi    turn     net
#>    <dbl> <chr>     <dbl>  <dbl>   <dbl>   <dbl>
#>  1  1990 BDI       61630 381611  443241 -319981
#>  2  1990 COM        9009  12011   21020   -3002
#>  3  1990 DJI       10949  55945   66894  -44996
#>  4  1990 ERI       14633 329383  344016 -314750
#>  5  1990 ETH     1635513 177334 1812847 1458179
#>  6  1990 KEN      306517  84833  391350  221684
#>  7  1990 MDG        9706  19159   28865   -9453
#>  8  1990 MWI      112416 974278 1086694 -861862
#>  9  1990 MUS       16862  22475   39337   -5613
#> 10  1990 MYT       13763   3021   16784   10742
#> # ... with 1,178 more rows

# }