Dictionary to look up region geographies based on countries used in UN DESA International Migrant Stock.

Intended for use as a custom dictionary with the countrycode package, where the existing UN region and area codes do not match those used by UN DESA in the WPP, see https://github.com/vincentarelbundock/countrycode/issues/253

Usage

dict_ims

Format

Data frame with 243 rows and 18 columns. One of first three columns intended as input for origin in countrycode.

name

Country name

iso3c

ISO numeric code

iso3n

ISO 3 letter code

Remaining columns intended as input for destination in countrycode.

name_short

Short country name

ims

Country in UN DESA International Migration Stock data. Some codes added for older political geographies to match World Bank data and older country units in IMS

region

Geographic region of country (6)

region_sub

Geographic sub region of country (22). Filled using region if none given in original data

region_sdg

SDG region of country (8)

region_sdg_sub

Sub SDG region of country (9). Filled using region_sdg if none given in original data

region_wb

World Bank region

un_develop

UN development group of country (3)

wb_income

World Bank income group of country (3)

wb_income_detail

Detailled World Bank income group of country (4)

lldc

Indicator variable for Land-Locked Developing Countries (32)

sids

Indicator variable for Small Island Developing States (58)

region_as2014

Region grouping used for global chord diagram plots by Abel and Sander (2014)

region_sab2014

Region grouping used for global chord diagram plots by Sander, Abel and Bauer (2014)

region_a2018

Region grouping used for global chord diagram plots by Abel (2018)

region_ac2022

Region grouping used for global chord diagram plots by Abel and Cohen (2022)

Source

The aggregates_correspondence_table_2020_1.xlsx file of United Nations Department of Economic and Social Affairs, Population Division (2020). International Migrant Stock 2020.

Examples

dict_ims
#> # A tibble: 243 × 18
#>    name      iso3c iso3n name_short ims   region region_sub region_sdg region_wb
#>    <chr>     <chr> <dbl> <chr>      <lgl> <chr>  <chr>      <chr>      <chr>    
#>  1 Burundi   BDI     108 Burundi    TRUE  Africa Eastern A… Sub-Sahar… Sub-Saha…
#>  2 Comoros   COM     174 Comoros    TRUE  Africa Eastern A… Sub-Sahar… Sub-Saha…
#>  3 Djibouti  DJI     262 Djibouti   TRUE  Africa Eastern A… Sub-Sahar… Middle E…
#>  4 Eritrea   ERI     232 Eritrea    TRUE  Africa Eastern A… Sub-Sahar… Sub-Saha…
#>  5 Ethiopia  ETH     231 Ethiopia   TRUE  Africa Eastern A… Sub-Sahar… Sub-Saha…
#>  6 Ethiopia… ETI     231 Ethiopia … FALSE Africa Eastern A… Sub-Sahar… Sub-Saha…
#>  7 Kenya     KEN     404 Kenya      TRUE  Africa Eastern A… Sub-Sahar… Sub-Saha…
#>  8 Madagasc… MDG     450 Madagascar TRUE  Africa Eastern A… Sub-Sahar… Sub-Saha…
#>  9 Malawi    MWI     454 Malawi     TRUE  Africa Eastern A… Sub-Sahar… Sub-Saha…
#> 10 Mauritius MUS     480 Mauritius  TRUE  Africa Eastern A… Sub-Sahar… Sub-Saha…
#> # ℹ 233 more rows
#> # ℹ 9 more variables: un_develop <chr>, wb_income <chr>,
#> #   wb_income_detail <chr>, lldc <lgl>, sids <lgl>, region_as2014 <chr>,
#> #   region_sab2014 <chr>, region_a2018 <chr>, region_ac2022 <chr>
# \dontrun{
library(tidyverse)
#> ── Attaching core tidyverse packages ──────────────────────── tidyverse 2.0.0 ──
#> ✔ dplyr     1.1.4     ✔ readr     2.1.5
#> ✔ forcats   1.0.0     ✔ stringr   1.5.1
#> ✔ ggplot2   3.5.1     ✔ tibble    3.2.1
#> ✔ lubridate 1.9.4     ✔ tidyr     1.3.1
#> ✔ purrr     1.0.2     
#> ── Conflicts ────────────────────────────────────────── tidyverse_conflicts() ──
#> ✖ dplyr::filter() masks stats::filter()
#> ✖ dplyr::lag()    masks stats::lag()
#> ℹ Use the conflicted package to force all conflicts to become errors
library(countrycode)
# download Abel and Cohen (2019) estimates
f <- read_csv("https://ndownloader.figshare.com/files/38016762", show_col_types = FALSE)
f
#> # A tibble: 307,833 × 9
#>    year0 orig  dest  sd_drop_neg sd_rev_neg mig_rate da_min_open da_min_closed
#>    <dbl> <chr> <chr>       <dbl>      <dbl>    <dbl>       <dbl>         <dbl>
#>  1  1990 BDI   BDI             0          0      0          0               0 
#>  2  1990 COM   BDI             0          0      0          0               0 
#>  3  1990 DJI   BDI             0          0      0          0               0 
#>  4  1990 ERI   BDI             0          0      0          0               0 
#>  5  1990 ETH   BDI             0          0      0          0               0 
#>  6  1990 KEN   BDI            30         30     75.7       51.3           207.
#>  7  1990 MDG   BDI             0          0      0          0.03            0 
#>  8  1990 MWI   BDI             0          0      0          0               0 
#>  9  1990 MUS   BDI             0          0      0          0.06            0 
#> 10  1990 MYT   BDI             0          0      0          0               0 
#> # ℹ 307,823 more rows
#> # ℹ 1 more variable: da_pb_closed <dbl>

# use dictionary to get region to region flows
d <- f %>%
  mutate(
    orig = countrycode(
      sourcevar = orig, custom_dict = dict_ims,
      origin = "iso3c", destination = "region"),
    dest = countrycode(
      sourcevar = dest, custom_dict = dict_ims,
      origin = "iso3c", destination = "region")
  ) %>%
  group_by(year0, orig, dest) %>%
  summarise_all(sum)
d
#> # A tibble: 216 × 9
#> # Groups:   year0, orig [36]
#>    year0 orig   dest   sd_drop_neg sd_rev_neg mig_rate da_min_open da_min_closed
#>    <dbl> <chr>  <chr>        <dbl>      <dbl>    <dbl>       <dbl>         <dbl>
#>  1  1990 Africa Africa     4297155    7845806   5.47e6    6872677.      7728373.
#>  2  1990 Africa Asia        240464     258816   7.24e5     283708.       554047.
#>  3  1990 Africa Europe      555826     664496   1.91e6     830461.      2190967.
#>  4  1990 Africa Latin…        1505       2709   7.81e3       9043.        56747.
#>  5  1990 Africa North…      289058     301706   2.23e5     321650.       783334.
#>  6  1990 Africa Ocean…       21550      23570   6.59e4      30186.       165598.
#>  7  1990 Asia   Africa       94088     158903   2.00e5     102036.        93577.
#>  8  1990 Asia   Asia       3616112    8617460   1.44e7    6969956.     10337980.
#>  9  1990 Asia   Europe     1496141    2322839   5.48e6    2851352.      4214903.
#> 10  1990 Asia   Latin…       14316      14343   1.07e5      20177.       136270.
#> # ℹ 206 more rows
#> # ℹ 1 more variable: da_pb_closed <dbl>
# }