Dictionary to look up region geographies based on countries used in UN DESA International Migrant Stock.

Intended for use as a custom dictionary with the countrycode package, where the existing UN region and area codes do not match those used by UN DESA in the WPP, see https://github.com/vincentarelbundock/countrycode/issues/253

dict_ims

Format

Data frame with 237 rows and 13 columns. One of first three columns intended as input for origin in countrycode.

name

Country name

iso3n

ISO 3 letter code

iso3c

ISO numeric code

Remaining columns intended as input for destination in countrycode.

region

Geographic region of country (6)

region_sub

Geographic sub region of country (22). Filled using region if none given in original data

region_sdg

SDG region of country (8)

region_sdg_sub

Sub SDG region of country (9). Filled using region_sdg if none given in original data

un_develop

UN development group of country (3)

wb_income

World Bank income group of country (3)

wb_income_detail

Detailled World Bank income group of country (4)

lldc

Indicator variable for Land-Locked Developing Countries (32)

sids

Indicator variable for Small Island Developing States (58)

region_as2014

Region grouping used for global chord diagram plots by Abel and Sander (2014)

region_sba2014

Region grouping used for global chord diagram plots by Sander, Abel and Bauer (2014)

region_a2018

Region grouping used for global chord diagram plots by Abel (2018)

region_ac2021

Region grouping used for global chord diagram plots by Abel and Cohen (2021)

Source

The aggregates_correspondence_table_2020_1.xlsx file of United Nations Department of Economic and Social Affairs, Population Division (2020). International Migrant Stock 2020.

Examples

# \dontrun{
library(tidyverse)

#> -- Attaching packages --------------------------------------- tidyverse 1.3.1 --

#> v ggplot2 3.3.5     v purrr   0.3.4
#> v tibble  3.1.5     v dplyr   1.0.7
#> v tidyr   1.1.4     v stringr 1.4.0
#> v readr   2.1.0     v forcats 0.5.1

#> Warning: package 'readr' was built under R version 4.1.2

#> -- Conflicts ------------------------------------------ tidyverse_conflicts() --
#> x dplyr::filter() masks stats::filter()
#> x dplyr::lag()    masks stats::lag()

library(countrycode)
# download Abel and Cohen (2019) estimates
f <- read_csv("https://ndownloader.figshare.com/files/26239945")

#> Rows: 235236 Columns: 9

#> -- Column specification --------------------------------------------------------
#> Delimiter: ","
#> chr (2): orig, dest
#> dbl (7): year0, sd_drop_neg, sd_rev_neg, mig_rate, da_min_open, da_min_close...

#> 
#> i Use `spec()` to retrieve the full column specification for this data.
#> i Specify the column types or set `show_col_types = FALSE` to quiet this message.

# use dictionary to get region to region flows
d <- f %>%
  mutate(
    orig = countrycode(
      sourcevar = orig, custom_dict = dict_ims,
      origin = "iso3c", destination = "region"),
    dest = countrycode(
      sourcevar = dest, custom_dict = dict_ims,
      origin = "iso3c", destination = "region")
  ) %>%
  group_by(year0, orig, dest) %>%
  summarise_all(sum)
d

#> # A tibble: 216 x 9
#> # Groups:   year0, orig [36]
#>    year0 orig   dest   sd_drop_neg sd_rev_neg mig_rate da_min_open da_min_closed
#>    <dbl> <chr>  <chr>        <dbl>      <dbl>    <dbl>       <dbl>         <dbl>
#>  1  1990 Africa Africa     4297155    7845712  4979307     6902972       7429916
#>  2  1990 Africa Asia        240464     258816   659002      287789        690446
#>  3  1990 Africa Europe      554285     662901  1734769      809487       1303377
#>  4  1990 Africa Latin~        1439       2624     7032        8452         12138
#>  5  1990 Africa North~      288992     301617   203364      318842        335302
#>  6  1990 Africa Ocean~       21541      23533    60019       27919         25460
#>  7  1990 Asia   Africa       94088     158903   182438      104503        318467
#>  8  1990 Asia   Asia       3616112    8617460 13154071     6839964       9534627
#>  9  1990 Asia   Europe     1495859    2322557  4987315     2855373       4465086
#> 10  1990 Asia   Latin~       14156      14183    97518       19285         44768
#> # ... with 206 more rows, and 1 more variable: da_pb_closed <dbl>

# }