Dictionary to look up region geographies based on countries used in UN DESA International Migrant Stock.

Intended for use as a custom dictionary with the countrycode package, where the existing UN region and area codes do not match those used by UN DESA in the WPP, see https://github.com/vincentarelbundock/countrycode/issues/253

dict_ims

Format

Data frame with 243 rows and 18 columns. One of first three columns intended as input for origin in countrycode.

name

Country name

iso3c

ISO numeric code

iso3n

ISO 3 letter code

Remaining columns intended as input for destination in countrycode.

name_short

Short country name

ims

Country in UN DESA International Migration Stock data. Some codes added for older political geographies to match World Bank data and older country units in IMS

region

Geographic region of country (6)

region_sub

Geographic sub region of country (22). Filled using region if none given in original data

region_sdg

SDG region of country (8)

region_sdg_sub

Sub SDG region of country (9). Filled using region_sdg if none given in original data

region_wb

World Bank region

un_develop

UN development group of country (3)

wb_income

World Bank income group of country (3)

wb_income_detail

Detailled World Bank income group of country (4)

lldc

Indicator variable for Land-Locked Developing Countries (32)

sids

Indicator variable for Small Island Developing States (58)

region_as2014

Region grouping used for global chord diagram plots by Abel and Sander (2014)

region_sab2014

Region grouping used for global chord diagram plots by Sander, Abel and Bauer (2014)

region_a2018

Region grouping used for global chord diagram plots by Abel (2018)

region_ac2022

Region grouping used for global chord diagram plots by Abel and Cohen (2022)

Source

The aggregates_correspondence_table_2020_1.xlsx file of United Nations Department of Economic and Social Affairs, Population Division (2020). International Migrant Stock 2020.

Examples

# \dontrun{
library(tidyverse)
#> ── Attaching core tidyverse packages ──────────────────────── tidyverse 2.0.0 ──
#> ✔ dplyr     1.1.0     ✔ readr     2.1.4
#> ✔ forcats   1.0.0     ✔ stringr   1.5.0
#> ✔ ggplot2   3.4.1     ✔ tibble    3.2.1
#> ✔ lubridate 1.9.2     ✔ tidyr     1.3.0
#> ✔ purrr     1.0.2     
#> ── Conflicts ────────────────────────────────────────── tidyverse_conflicts() ──
#> ✖ dplyr::filter() masks stats::filter()
#> ✖ dplyr::lag()    masks stats::lag()
#> ℹ Use the conflicted package (<http://conflicted.r-lib.org/>) to force all conflicts to become errors
library(countrycode)
# download Abel and Cohen (2019) estimates
f <- read_csv("https://ndownloader.figshare.com/files/38016762", show_col_types = FALSE)
f
#> # A tibble: 307,833 × 9
#>    year0 orig  dest  sd_drop_neg sd_rev_neg mig_rate da_min_open da_mi…¹ da_pb…²
#>    <dbl> <chr> <chr>       <dbl>      <dbl>    <dbl>       <dbl>   <dbl>   <dbl>
#>  1  1990 BDI   BDI             0          0      0          0         0     0   
#>  2  1990 COM   BDI             0          0      0          0         0     0   
#>  3  1990 DJI   BDI             0          0      0          0         0     0   
#>  4  1990 ERI   BDI             0          0      0          0         0    38.3 
#>  5  1990 ETH   BDI             0          0      0          0         0     0.82
#>  6  1990 KEN   BDI            30         30     75.7       51.3     207.  248.  
#>  7  1990 MDG   BDI             0          0      0          0.03      0     0   
#>  8  1990 MWI   BDI             0          0      0          0         0    98.8 
#>  9  1990 MUS   BDI             0          0      0          0.06      0     0   
#> 10  1990 MYT   BDI             0          0      0          0         0     0   
#> # … with 307,823 more rows, and abbreviated variable names ¹​da_min_closed,
#> #   ²​da_pb_closed

# use dictionary to get region to region flows
d <- f %>%
  mutate(
    orig = countrycode(
      sourcevar = orig, custom_dict = dict_ims,
      origin = "iso3c", destination = "region"),
    dest = countrycode(
      sourcevar = dest, custom_dict = dict_ims,
      origin = "iso3c", destination = "region")
  ) %>%
  group_by(year0, orig, dest) %>%
  summarise_all(sum)
d
#> # A tibble: 216 × 9
#> # Groups:   year0, orig [36]
#>    year0 orig   dest             sd_dr…¹ sd_re…² mig_r…³ da_mi…⁴ da_mi…⁵ da_pb…⁶
#>    <dbl> <chr>  <chr>              <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>
#>  1  1990 Africa Africa           4297155 7845806  5.47e6  6.87e6  7.73e6  1.00e7
#>  2  1990 Africa Asia              240464  258816  7.24e5  2.84e5  5.54e5  8.15e5
#>  3  1990 Africa Europe            555826  664496  1.91e6  8.30e5  2.19e6  2.74e6
#>  4  1990 Africa Latin America a…    1505    2709  7.81e3  9.04e3  5.67e4  5.46e4
#>  5  1990 Africa Northern America  289058  301706  2.23e5  3.22e5  7.83e5  8.28e5
#>  6  1990 Africa Oceania            21550   23570  6.59e4  3.02e4  1.66e5  1.77e5
#>  7  1990 Asia   Africa             94088  158903  2.00e5  1.02e5  9.36e4  3.53e5
#>  8  1990 Asia   Asia             3616112 8617460  1.44e7  6.97e6  1.03e7  1.71e7
#>  9  1990 Asia   Europe           1496141 2322839  5.48e6  2.85e6  4.21e6  6.42e6
#> 10  1990 Asia   Latin America a…   14316   14343  1.07e5  2.02e4  1.36e5  1.75e5
#> # … with 206 more rows, and abbreviated variable names ¹​sd_drop_neg,
#> #   ²​sd_rev_neg, ³​mig_rate, ⁴​da_min_open, ⁵​da_min_closed, ⁶​da_pb_closed
# }