Estimation of bilateral migrant flows from bilateral migrant stocks using demographic accounting approaches

Estimates migrant transitions flows between two sequential migrant stock tables. Replaces old ffs.

ffs_demo(
  m1 = NULL,
  m2 = NULL,
  b_por = NULL,
  d_por = NULL,
  m = NULL,
  stayer_assumption = TRUE,
  match_global = "before-demo-adjust",
  match_pob_tot_method = "rescale",
  birth_non_negative = TRUE,
  death_method = "proportion",
  verbose = FALSE,
  ...
)

Arguments

m1
Matrix of migrant stock totals at time t. Rows in the matrix correspond to place of birth and columns to place of residence at time t

m2
Matrix of migrant stock totals at time t+1. Rows in the matrix correspond to place of birth and columns to place of residence at time t+1.

b_por
Vector of the number of births between time t and t+1 in each region.

d_por
Vector of the number of deaths between time t and t+1 in each region.

m
Matrix of auxiliary data. By default set to 1 for all origin-destination combinations.

stayer_assumption
Logical value to indicate whether to use ipf3 or ipf3_qi to estimate flows. By default uses ipf3_qi, i.e. is set to TRUE. The ipf function is useful for replicating method of Azoze and Raferty.

match_global
Character string used to indicate whether to balance the change in stocks totals with the changes in births and deaths. Only applied when match_pob_tot_method is either rescale or rescale-adjust-zero-fb. By default uses after-demo-adjust rather than before-demo-adjust which I think minimises risk of negative values.

match_pob_tot_method
Character string passed to method argument in match_pob_tot to ensure place of birth margins in stock tables match.

birth_non_negative
Logical value passed to non_negative argument in birth_mat.

death_method
Character string passed to method argument in death_mat.

verbose
Logical value to indicate the print the parameter estimates at each iteration of the various IPF routines. By default FALSE.

...
Additional arguments passes to ipf3_qi or ipf3.

Value

Estimates migrant transitions flows between two sequential migrant stock tables using various methods. See the example section for possible variations on estimation methods.

Returns a list object with:

mu

Array of indirect estimates of origin-destination matrices by place of birth.

it

Iteration count.

tol

Tolerance level at final iteration.

y

Array of indirect estimates of origin-destination matrices by place of birth with additional rows and columns for births, deaths and moves to other regions.

...

Slots to record which estimation method was used (as set by arguments above)

od_flow

Matrix of estimated origin-destination flows

References

Abel and Cohen (2019) Bilateral international migration flow estimates for 200 countries Scientific Data 6 (1), 1-13

Azose & Raftery (2019) Estimation of emigration, return migration, and transit migration between all pairs of countries Proceedings of the National Academy of Sciences 116 (1) 116-122

Abel, G. J. (2018). Estimates of Global Bilateral Migration Flows by Gender between 1960 and 2015. International Migration Review 52 (3), 809–852.

Abel, G. J. and Sander, N. (2014). Quantifying Global International Migration Flows. Science, 343 (6178) 1520-1522

Abel, G. J. (2013). Estimating Global Migration Flow Tables Using Place of Birth. Demographic Research 28, (18) 505-546

See also

Author

Guy J. Abel

Examples

##
## without births and deaths over period
##
# data as in papers
s1 <- matrix(data = c(1000, 100, 10, 0, 55, 555, 50, 5, 80, 40, 800, 40, 20, 25, 20, 200),
             nrow = 4, ncol = 4, byrow = TRUE)
s2 <- matrix(data = c(950, 100, 60, 0, 80, 505, 75, 5, 90, 30, 800, 40, 40, 45, 0, 180),
             nrow = 4, ncol = 4, byrow = TRUE)
b <- d <- rep(0, 4)
reg <- LETTERS[1:4]
dimnames(s1) <- dimnames(s2) <- list(pob = reg, por = reg)
names(b) <- names(d) <- reg
s1; s2; b; d

#>    por
#> pob    A   B   C   D
#>   A 1000 100  10   0
#>   B   55 555  50   5
#>   C   80  40 800  40
#>   D   20  25  20 200

#>    por
#> pob   A   B   C   D
#>   A 950 100  60   0
#>   B  80 505  75   5
#>   C  90  30 800  40
#>   D  40  45   0 180

#> A B C D 
#> 0 0 0 0 

#> A B C D 
#> 0 0 0 0 

# demographic research and science paper example
ffs_demo(m1 = s1, m2 = s2, b_por = b, d_por = d)

#> $flow
#> , , pob = A
#> 
#>     dest
#> orig     A   B  C D
#>    A 950.5   0 50 0
#>    B   0.0 100  0 0
#>    C   0.0   0 10 0
#>    D   0.0   0  0 0
#> 
#> , , pob = B
#> 
#>     dest
#> orig  A   B  C D
#>    A 55   0  0 0
#>    B 25 505 25 0
#>    C  0   0 50 0
#>    D  0   0  0 5
#> 
#> , , pob = C
#> 
#>     dest
#> orig  A  B   C  D
#>    A 80  0   0  0
#>    B 10 30   0  0
#>    C  0  0 800  0
#>    D  0  0   0 40
#> 
#> , , pob = D
#> 
#>     dest
#> orig  A  B C   D
#>    A 20  0 0   0
#>    B  0 25 0   0
#>    C 10 10 0   0
#>    D 10 10 0 180
#> 
#> 
#> $y
#> , , pob = A
#> 
#>          dest
#> orig          A   B  C D death outside
#>   A       950.5   0 50 0     0       0
#>   B         0.0 100  0 0     0       0
#>   C         0.0   0 10 0     0       0
#>   D         0.0   0  0 0     0       0
#>   birth     0.0   0  0 0     0       0
#>   outside   0.0   0  0 0     0       0
#> 
#> , , pob = B
#> 
#>          dest
#> orig       A   B  C D death outside
#>   A       55   0  0 0     0       0
#>   B       25 505 25 0     0       0
#>   C        0   0 50 0     0       0
#>   D        0   0  0 5     0       0
#>   birth    0   0  0 0     0       0
#>   outside  0   0  0 0     0       0
#> 
#> , , pob = C
#> 
#>          dest
#> orig       A  B   C  D death outside
#>   A       80  0   0  0     0       0
#>   B       10 30   0  0     0       0
#>   C        0  0 800  0     0       0
#>   D        0  0   0 40     0       0
#>   birth    0  0   0  0     0       0
#>   outside  0  0   0  0     0       0
#> 
#> , , pob = D
#> 
#>          dest
#> orig       A  B C   D death outside
#>   A       20  0 0   0     0       0
#>   B        0 25 0   0     0       0
#>   C       10 10 0   0     0       0
#>   D       10 10 0 180     0       0
#>   birth    0  0 0   0     0       0
#>   outside  0  0 0   0     0       0
#> 
#> 
#> $stayer_assumption
#> [1] TRUE
#> 
#> $match_pob_tot_method
#> [1] "rescale"
#> 
#> $birth_non_negative
#> [1] TRUE
#> 
#> $death_method
#> [1] "proportion"
#> 
#> $od_flow
#>      dest
#> orig   A  B  C D Sum
#>   A    0  0 50 0  50
#>   B   35  0 25 0  60
#>   C   10 10  0 0  20
#>   D   10 10  0 0  20
#>   Sum 55 20 75 0 150
#> 

# international migration review paper example
s1[,] <- c(100, 20, 10, 20, 10, 55, 40, 25, 10, 25, 140, 20, 0, 10, 65, 200)
s2[,] <- c(70, 25, 10, 40, 30, 60, 55, 45, 10, 10, 140, 0, 10, 15, 50, 180)
ffs_demo(m1 = s1, m2 = s2, b_por = b, d_por = d)

#> $flow
#> , , pob = A
#> 
#>     dest
#> orig        A        B          C        D
#>    A 69.97555 19.99853 0.04130057 9.998155
#>    B  0.00000 10.00214 0.00000000 0.000000
#>    C  0.00000  0.00000 9.98432596 0.000000
#>    D  0.00000  0.00000 0.00000000 0.000000
#> 
#> , , pob = B
#> 
#>     dest
#> orig         A         B      C        D
#>    A 19.961779  0.000000  0.000 0.000000
#>    B  0.000000 55.398318  0.000 0.000000
#>    C  5.082765  4.730935 10.047 5.049044
#>    D  0.000000  0.000000  0.000 9.980160
#> 
#> , , pob = C
#> 
#>     dest
#> orig        A           B            C        D
#>    A 9.963367  0.06259027 3.087515e-03  0.00000
#>    B 0.000000 40.11935070 0.000000e+00  0.00000
#>    C 0.000000  0.00000000 1.401676e+02  0.00000
#>    D 0.000000 14.63694990 7.220260e-01 49.82505
#> 
#> , , pob = D
#> 
#>     dest
#> orig         A         B C        D
#>    A 19.984653  0.000000 0   0.0000
#>    B  0.000000 24.982793 0   0.0000
#>    C  9.969731  9.980899 0   0.0000
#>    D 10.076411 10.087698 0 179.6678
#> 
#> 
#> $y
#> , , pob = A
#> 
#>          dest
#> orig             A        B          C        D death outside
#>   A       69.97555 19.99853 0.04130057 9.998155     0       0
#>   B        0.00000 10.00214 0.00000000 0.000000     0       0
#>   C        0.00000  0.00000 9.98432596 0.000000     0       0
#>   D        0.00000  0.00000 0.00000000 0.000000     0       0
#>   birth    0.00000  0.00000 0.00000000 0.000000     0       0
#>   outside  0.00000  0.00000 0.00000000 0.000000     0       0
#> 
#> , , pob = B
#> 
#>          dest
#> orig              A         B      C        D death outside
#>   A       19.961779  0.000000  0.000 0.000000     0       0
#>   B        0.000000 55.398318  0.000 0.000000     0       0
#>   C        5.082765  4.730935 10.047 5.049044     0       0
#>   D        0.000000  0.000000  0.000 9.980160     0       0
#>   birth    0.000000  0.000000  0.000 0.000000     0       0
#>   outside  0.000000  0.000000  0.000 0.000000     0       0
#> 
#> , , pob = C
#> 
#>          dest
#> orig             A           B            C        D death outside
#>   A       9.963367  0.06259027 3.087515e-03  0.00000     0       0
#>   B       0.000000 40.11935070 0.000000e+00  0.00000     0       0
#>   C       0.000000  0.00000000 1.401676e+02  0.00000     0       0
#>   D       0.000000 14.63694990 7.220260e-01 49.82505     0       0
#>   birth   0.000000  0.00000000 0.000000e+00  0.00000     0       0
#>   outside 0.000000  0.00000000 0.000000e+00  0.00000     0       0
#> 
#> , , pob = D
#> 
#>          dest
#> orig              A         B C        D death outside
#>   A       19.984653  0.000000 0   0.0000     0       0
#>   B        0.000000 24.982793 0   0.0000     0       0
#>   C        9.969731  9.980899 0   0.0000     0       0
#>   D       10.076411 10.087698 0 179.6678     0       0
#>   birth    0.000000  0.000000 0   0.0000     0       0
#>   outside  0.000000  0.000000 0   0.0000     0       0
#> 
#> 
#> $stayer_assumption
#> [1] TRUE
#> 
#> $match_pob_tot_method
#> [1] "rescale"
#> 
#> $birth_non_negative
#> [1] TRUE
#> 
#> $death_method
#> [1] "proportion"
#> 
#> $od_flow
#>      dest
#> orig         A        B          C         D       Sum
#>   A    0.00000 20.06112 0.04438808  9.998155  30.10366
#>   B    0.00000  0.00000 0.00000000  0.000000   0.00000
#>   C   15.05250 14.71183 0.00000000  5.049044  34.81337
#>   D   10.07641 24.72465 0.72202599  0.000000  35.52308
#>   Sum 25.12891 59.49760 0.76641407 15.047199 100.44012
#> 

# international migration review supp. material example
dm <- matrix(data = c(0, 5, 50, 500, 5, 0, 45, 495, 50, 45, 0, 450, 500, 495, 450, 0), 
             nrow = 4, ncol = 4, byrow = TRUE)
dimnames(dm) <- list(orig = reg, dest = reg)
ffs_demo(m1 = s1, m2 = s2, b_por = b, d_por = d, m = dm)

#> $flow
#> , , pob = A
#> 
#>     dest
#> orig        A           B        C           D
#>    A 70.37889 19.81927767 0.000000 9.973302073
#>    B  0.00000 10.04225735 0.000000 0.000000000
#>    C  0.00000  0.05684745 9.976561 0.002860633
#>    D  0.00000  0.00000000 0.000000 0.000000000
#> 
#> , , pob = B
#> 
#>     dest
#> orig         A         B        C        D
#>    A 19.917879  0.000000  0.00000 0.000000
#>    B  0.000000 55.410664  0.00000 0.000000
#>    C  5.156388  4.707128 10.02342 5.057943
#>    D  0.000000  0.000000  0.00000 9.976581
#> 
#> , , pob = C
#> 
#>     dest
#> orig          A        B            C        D
#>    A 9.98110783  0.00000 0.000000e+00  0.00000
#>    B 0.00000000 40.02451 0.000000e+00  0.00000
#>    C 0.00000000  0.00000 1.400023e+02  0.00000
#>    D 0.02581451 14.95828 6.799172e-03 50.00124
#> 
#> , , pob = D
#> 
#>     dest
#> orig         A         B C        D
#>    A 19.965709  0.000000 0   0.0000
#>    B  0.000000 25.019696 0   0.0000
#>    C 10.264289  9.739533 0   0.0000
#>    D  9.794964 10.223623 0 179.9922
#> 
#> 
#> $y
#> , , pob = A
#> 
#>          dest
#> orig             A           B        C           D death outside
#>   A       70.37889 19.81927767 0.000000 9.973302073     0       0
#>   B        0.00000 10.04225735 0.000000 0.000000000     0       0
#>   C        0.00000  0.05684745 9.976561 0.002860633     0       0
#>   D        0.00000  0.00000000 0.000000 0.000000000     0       0
#>   birth    0.00000  0.00000000 0.000000 0.000000000     0       0
#>   outside  0.00000  0.00000000 0.000000 0.000000000     0       0
#> 
#> , , pob = B
#> 
#>          dest
#> orig              A         B        C        D death outside
#>   A       19.917879  0.000000  0.00000 0.000000     0       0
#>   B        0.000000 55.410664  0.00000 0.000000     0       0
#>   C        5.156388  4.707128 10.02342 5.057943     0       0
#>   D        0.000000  0.000000  0.00000 9.976581     0       0
#>   birth    0.000000  0.000000  0.00000 0.000000     0       0
#>   outside  0.000000  0.000000  0.00000 0.000000     0       0
#> 
#> , , pob = C
#> 
#>          dest
#> orig               A        B            C        D death outside
#>   A       9.98110783  0.00000 0.000000e+00  0.00000     0       0
#>   B       0.00000000 40.02451 0.000000e+00  0.00000     0       0
#>   C       0.00000000  0.00000 1.400023e+02  0.00000     0       0
#>   D       0.02581451 14.95828 6.799172e-03 50.00124     0       0
#>   birth   0.00000000  0.00000 0.000000e+00  0.00000     0       0
#>   outside 0.00000000  0.00000 0.000000e+00  0.00000     0       0
#> 
#> , , pob = D
#> 
#>          dest
#> orig              A         B C        D death outside
#>   A       19.965709  0.000000 0   0.0000     0       0
#>   B        0.000000 25.019696 0   0.0000     0       0
#>   C       10.264289  9.739533 0   0.0000     0       0
#>   D        9.794964 10.223623 0 179.9922     0       0
#>   birth    0.000000  0.000000 0   0.0000     0       0
#>   outside  0.000000  0.000000 0   0.0000     0       0
#> 
#> 
#> $stayer_assumption
#> [1] TRUE
#> 
#> $match_pob_tot_method
#> [1] "rescale"
#> 
#> $birth_non_negative
#> [1] TRUE
#> 
#> $death_method
#> [1] "proportion"
#> 
#> $od_flow
#>      dest
#> orig          A        B           C         D      Sum
#>   A    0.000000 19.81928 0.000000000  9.973302 29.79258
#>   B    0.000000  0.00000 0.000000000  0.000000  0.00000
#>   C   15.420677 14.50351 0.000000000  5.060804 34.98499
#>   D    9.820779 25.18190 0.006799172  0.000000 35.00948
#>   Sum 25.241456 59.50469 0.006799172 15.034106 99.78705
#> 

##
## with births and deaths over period
##
# demographic research paper example
s1[,] <- c(1000, 55, 80, 20, 100, 555, 40, 25, 10, 50, 800, 20, 0, 5, 40, 200)
s2[,] <- c(1060, 45, 70, 30, 60, 540, 75, 30, 10, 40, 770, 20, 10, 0, 70, 230)
b[] <- c(80, 20, 40, 60)
d[] <- c(70, 30, 50, 10)
ffs_demo(m1 = s1, m2 = s2, b_por = b, d_por = d, match_pob_tot_method = "open-dr")

#> $flow
#> , , pob = A
#> 
#>     dest
#> orig         A        B         C        D
#>    A 939.39394  0.00000 0.0000000 0.000000
#>    B  26.42296 59.13165 0.4234563 9.855274
#>    C   0.00000  0.00000 9.4318182 0.000000
#>    D   0.00000  0.00000 0.0000000 0.000000
#> 
#> , , pob = B
#> 
#>     dest
#> orig  A          B  C D
#>    A 45   4.187249  0 0
#>    B  0 506.350996  0 0
#>    C  0   4.895986 40 0
#>    D  0   4.565768  0 0
#> 
#> , , pob = C
#> 
#>     dest
#> orig        A         B        C         D
#>    A 67.14057  4.315617   0.0000  3.695332
#>    B  0.00000 38.333333   0.0000  0.000000
#>    C  0.00000 29.287371 700.1802 25.077888
#>    D  0.00000  0.000000   0.0000 38.367347
#> 
#> , , pob = D
#> 
#>     dest
#> orig        A         B         C   D
#>    A 18.53239  0.000000  0.000000   0
#>    B  0.00000 23.632527  0.000000   0
#>    C  0.00000  0.000000 18.607113   0
#>    D 11.46761  6.367473  1.392887 170
#> 
#> 
#> $y
#> , , pob = A
#> 
#>          dest
#> orig              A        B         C        D      death    outside
#>   A       939.39394  0.00000 0.0000000 0.000000 60.6060606 14.1831046
#>   B        26.42296 59.13165 0.4234563 9.855274  4.1666667  0.8683533
#>   C         0.00000  0.00000 9.4318182 0.000000  0.5681818  0.1447256
#>   D         0.00000  0.00000 0.0000000 0.000000  0.0000000  0.1447256
#>   birth    80.00000  0.00000 0.0000000 0.000000  0.0000000  0.0000000
#>   outside   0.00000  0.00000 0.0000000 0.000000  0.0000000  0.0000000
#> 
#> , , pob = B
#> 
#>          dest
#> orig              A          B         C       D      death outside
#>   A       45.000000   4.187249  0.000000 0.00000  3.3333333       0
#>   B        0.000000 506.350996  0.000000 0.00000 23.1250000       0
#>   C        0.000000   4.895986 40.000000 0.00000  2.8409091       0
#>   D        0.000000   4.565768  0.000000 0.00000  0.2040816       0
#>   birth    0.000000  20.000000  0.000000 0.00000  0.0000000       0
#>   outside  2.479418  25.524004  2.263105 0.23015  0.0000000       0
#> 
#> , , pob = C
#> 
#>          dest
#> orig             A         B        C         D     death   outside
#>   A       67.14057  4.315617   0.0000  3.695332  4.848485  2.859433
#>   B        0.00000 38.333333   0.0000  0.000000  1.666667  3.063679
#>   C        0.00000 29.287371 700.1802 25.077888 45.454545 29.819805
#>   D        0.00000  0.000000   0.0000 38.367347  1.632653  2.859433
#>   birth    0.00000  0.000000  40.0000  0.000000  0.000000  0.000000
#>   outside  0.00000  0.000000   0.0000  0.000000  0.000000  0.000000
#> 
#> , , pob = D
#> 
#>          dest
#> orig               A          B          C         D    death outside
#>   A       18.5323852  0.0000000  0.0000000   0.00000 1.212121       0
#>   B        0.0000000 23.6325274  0.0000000   0.00000 1.041667       0
#>   C        0.0000000  0.0000000 18.6071125   0.00000 1.136364       0
#>   D       11.4676148  6.3674726  1.3928875 170.00000 8.163265       0
#>   birth    0.0000000  0.0000000  0.0000000  60.00000 0.000000       0
#>   outside  0.2554936  0.3258059  0.2565238   2.60876 0.000000       0
#> 
#> 
#> $stayer_assumption
#> [1] TRUE
#> 
#> $match_pob_tot_method
#> [1] "open-dr"
#> 
#> $birth_non_negative
#> [1] TRUE
#> 
#> $death_method
#> [1] "proportion"
#> 
#> $od_flow
#>      dest
#> orig         A         B         C         D       Sum
#>   A    0.00000  8.502866 0.0000000  3.695332  12.19820
#>   B   26.42296  0.000000 0.4234563  9.855274  36.70169
#>   C    0.00000 34.183358 0.0000000 25.077888  59.26125
#>   D   11.46761 10.933241 1.3928875  0.000000  23.79374
#>   Sum 37.89057 53.619464 1.8163437 38.628494 131.95487
#> 

# makes more sense to use this method
ffs_demo(m1 = s1, m2 = s2, b_por = b, d_por = d, match_pob_tot_method = "open")

#> $flow
#> , , pob = A
#> 
#>     dest
#> orig         A        B         C  D
#>    A 939.39394  0.00000 0.0000000  0
#>    B  26.81098 58.59268 0.4296748 10
#>    C   0.00000  0.00000 9.4318182  0
#>    D   0.00000  0.00000 0.0000000  0
#> 
#> , , pob = B
#> 
#>     dest
#> orig  A          B  C D
#>    A 45   4.398319  0 0
#>    B  0 505.662981  0 0
#>    C  0   5.142782 40 0
#>    D  0   4.795918  0 0
#> 
#> , , pob = C
#> 
#>     dest
#> orig        A         B        C         D
#>    A 66.79939  4.499413   0.0000  3.852711
#>    B  0.00000 38.333333   0.0000  0.000000
#>    C  0.00000 30.534684 697.8648 26.145924
#>    D  0.00000  0.000000   0.0000 38.367347
#> 
#> , , pob = D
#> 
#>     dest
#> orig        A         B        C   D
#>    A 18.37429  0.000000  0.00000   0
#>    B  0.00000 23.544743  0.00000   0
#>    C  0.00000  0.000000 18.58791   0
#>    D 11.62571  6.455257  1.41209 170
#> 
#> 
#> $y
#> , , pob = A
#> 
#>          dest
#> orig              A        B         C  D      death    outside
#>   A       939.39394  0.00000 0.0000000  0 60.6060606 13.7950812
#>   B        26.81098 58.59268 0.4296748 10  4.1666667  1.4073208
#>   C         0.00000  0.00000 9.4318182  0  0.5681818  0.1385071
#>   D         0.00000  0.00000 0.0000000  0  0.0000000  0.0000000
#>   birth    80.00000  0.00000 0.0000000  0  0.0000000  0.0000000
#>   outside   0.00000  0.00000 0.0000000  0  0.0000000  0.0000000
#> 
#> , , pob = B
#> 
#>          dest
#> orig              A          B         C D      death outside
#>   A       45.000000   4.398319  0.000000 0  3.3333333       0
#>   B        0.000000 505.662981  0.000000 0 23.1250000       0
#>   C        0.000000   5.142782 40.000000 0  2.8409091       0
#>   D        0.000000   4.795918  0.000000 0  0.2040816       0
#>   birth    0.000000  20.000000  0.000000 0  0.0000000       0
#>   outside  2.268348  26.212019  2.016309 0  0.0000000       0
#> 
#> , , pob = C
#> 
#>          dest
#> orig             A         B        C         D     death   outside
#>   A       66.79939  4.499413   0.0000  3.852711  4.848485  3.200610
#>   B        0.00000 38.333333   0.0000  0.000000  1.666667  1.632569
#>   C        0.00000 30.534684 697.8648 26.145924 45.454545 32.135154
#>   D        0.00000  0.000000   0.0000 38.367347  1.632653  1.634018
#>   birth    0.00000  0.000000  40.0000  0.000000  0.000000  0.000000
#>   outside  0.00000  0.000000   0.0000  0.000000  0.000000  0.000000
#> 
#> , , pob = D
#> 
#>          dest
#> orig             A         B          C          D    death outside
#>   A       18.37429  0.000000  0.0000000   0.000000 1.212121       0
#>   B        0.00000 23.544743  0.0000000   0.000000 1.041667       0
#>   C        0.00000  0.000000 18.5879097   0.000000 1.136364       0
#>   D       11.62571  6.455257  1.4120903 170.000000 8.163265       0
#>   birth    0.00000  0.000000  0.0000000  60.000000 0.000000       0
#>   outside  0.41359  0.413590  0.2757267   2.343677 0.000000       0
#> 
#> 
#> $stayer_assumption
#> [1] TRUE
#> 
#> $match_pob_tot_method
#> [1] "open"
#> 
#> $birth_non_negative
#> [1] TRUE
#> 
#> $death_method
#> [1] "proportion"
#> 
#> $od_flow
#>      dest
#> orig         A         B         C         D       Sum
#>   A    0.00000  8.897732 0.0000000  3.852711  12.75044
#>   B   26.81098  0.000000 0.4296748 10.000000  37.24065
#>   C    0.00000 35.677466 0.0000000 26.145924  61.82339
#>   D   11.62571 11.251175 1.4120903  0.000000  24.28898
#>   Sum 38.43669 55.826373 1.8417650 39.998635 136.10346
#> 

# science paper  supp. material example
b[] <- c(80, 20, 60, 60)
ffs_demo(m1 = s1, m2 = s2, b_por = b, d_por = d)

#> $flow
#> , , pob = A
#> 
#>     dest
#> orig         A        B        C       D
#>    A 942.46268  0.00000 0.000000 0.00000
#>    B  34.10724 56.49634 0.583247 9.81133
#>    C   0.00000  0.00000 9.366678 0.00000
#>    D   0.00000  0.00000 0.000000 0.00000
#> 
#> , , pob = B
#> 
#>     dest
#> orig            A          B        C D
#>    A 48.389330144   0.000000  0.00000 0
#>    B  0.000000000 523.535673  0.00000 0
#>    C  0.002843014   0.751389 42.98745 0
#>    D  0.017272875   4.565102  0.00000 0
#> 
#> , , pob = C
#> 
#>     dest
#> orig       A         B        C         D
#>    A 69.7374  3.327696   0.0000  3.240618
#>    B  0.0000 40.906604   0.0000  0.000000
#>    C  0.0000 26.404335 706.6268 25.713389
#>    D  0.0000  0.000000   0.0000 39.743008
#> 
#> , , pob = D
#> 
#>     dest
#> orig        A         B         C        D
#>    A 18.30537  0.000000  0.000000   0.0000
#>    B  0.00000 24.533233  0.000000   0.0000
#>    C  0.00000  0.000000 18.201909   0.0000
#>    D 12.41690  4.511449  2.259055 171.4956
#> 
#> 
#> $y
#> , , pob = A
#> 
#>          dest
#> orig              A        B        C       D      death outside
#>   A       942.46268  0.00000 0.000000 0.00000 60.6101255       0
#>   B        34.10724 56.49634 0.583247 9.81133  4.1666667       0
#>   C         0.00000  0.00000 9.366678 0.00000  0.5681818       0
#>   D         0.00000  0.00000 0.000000 0.00000  0.0000000       0
#>   birth    80.00000  0.00000 0.000000 0.00000  0.0000000       0
#>   outside   0.00000  0.00000 0.000000 0.00000  0.0000000       0
#> 
#> , , pob = B
#> 
#>          dest
#> orig                 A          B        C D      death outside
#>   A       48.389330144   0.000000  0.00000 0  3.3318910       0
#>   B        0.000000000 523.535673  0.00000 0 23.1250000       0
#>   C        0.002843014   0.751389 42.98745 0  2.8409091       0
#>   D        0.017272875   4.565102  0.00000 0  0.2040816       0
#>   birth    0.000000000  20.000000  0.00000 0  0.0000000       0
#>   outside  0.000000000   0.000000  0.00000 0  0.0000000       0
#> 
#> , , pob = C
#> 
#>          dest
#> orig            A         B        C         D     death outside
#>   A       69.7374  3.327696   0.0000  3.240618  4.846387       0
#>   B        0.0000 40.906604   0.0000  0.000000  1.666667       0
#>   C        0.0000 26.404335 706.6268 25.713389 45.454545       0
#>   D        0.0000  0.000000   0.0000 39.743008  1.632653       0
#>   birth    0.0000  0.000000  60.0000  0.000000  0.000000       0
#>   outside  0.0000  0.000000   0.0000  0.000000  0.000000       0
#> 
#> , , pob = D
#> 
#>          dest
#> orig             A         B         C        D    death outside
#>   A       18.30537  0.000000  0.000000   0.0000 1.211597       0
#>   B        0.00000 24.533233  0.000000   0.0000 1.041667       0
#>   C        0.00000  0.000000 18.201909   0.0000 1.136364       0
#>   D       12.41690  4.511449  2.259055 171.4956 8.163265       0
#>   birth    0.00000  0.000000  0.000000  60.0000 0.000000       0
#>   outside  0.00000  0.000000  0.000000   0.0000 0.000000       0
#> 
#> 
#> $stayer_assumption
#> [1] TRUE
#> 
#> $match_pob_tot_method
#> [1] "rescale"
#> 
#> $birth_non_negative
#> [1] TRUE
#> 
#> $death_method
#> [1] "proportion"
#> 
#> $od_flow
#>      dest
#> orig             A         B        C         D        Sum
#>   A    0.000000000  3.327696 0.000000  3.240618   6.568314
#>   B   34.107236105  0.000000 0.583247  9.811330  44.501813
#>   C    0.002843014 27.155724 0.000000 25.713389  52.871956
#>   D   12.434176193  9.076551 2.259055  0.000000  23.769782
#>   Sum 46.544255312 39.559971 2.842302 38.765337 127.711866
#> 

# international migration review supp. material example
s1[,] <- c(100, 20, 10, 20, 10, 55, 40, 25, 10, 25, 140, 20, 0, 10, 65, 200)
s2[,] <- c(75, 20, 30, 30, 25, 45, 40, 30, 5, 30, 150, 20, 0, 15, 60, 230)
b[] <- c(10, 50, 25, 60)
d[] <- c(30, 10, 40, 10)
ffs_demo(m1 = s1, m2 = s2, b_por = b, d_por = d)

#> $flow
#> , , pob = A
#> 
#>     dest
#> orig        A        B        C D
#>    A 68.73684 8.258457 0.000000 0
#>    B  0.00000 8.120586 0.000000 0
#>    C  0.00000 1.051386 5.675578 0
#>    D  0.00000 0.000000 0.000000 0
#> 
#> , , pob = B
#> 
#>     dest
#> orig         A        B        C        D
#>    A 17.058087  0.00000  0.00000 0.000000
#>    B  2.495671 28.68482 12.76128 5.780648
#>    C  0.000000  0.00000 18.72239 0.000000
#>    D  0.000000  0.00000  0.00000 9.417525
#> 
#> , , pob = C
#> 
#>     dest
#> orig         A        B          C        D
#>    A  9.318430  0.00000   0.000000  0.00000
#>    B 10.706236 23.92586   4.876536  0.00000
#>    C  0.000000  0.00000 114.549126  0.00000
#>    D  7.192126  0.00000   3.275910 56.41139
#> 
#> , , pob = D
#> 
#>     dest
#> orig         A        B         C        D
#>    A 17.092733  0.00000  0.000000   0.0000
#>    B  2.132977 19.46418  1.049853   0.0000
#>    C  0.000000  0.00000 15.008335   0.0000
#>    D 10.296176  0.00000  5.067787 173.3691
#> 
#> 
#> $y
#> , , pob = A
#> 
#>          dest
#> orig             A        B        C D      death outside
#>   A       68.73684 8.258457 0.000000 0 20.0332226       0
#>   B        0.00000 8.120586 0.000000 0  0.7692308       0
#>   C        0.00000 1.051386 5.675578 0  2.0512821       0
#>   D        0.00000 0.000000 0.000000 0  0.0000000       0
#>   birth   10.00000 0.000000 0.000000 0  0.0000000       0
#>   outside  0.00000 0.000000 0.000000 0  0.0000000       0
#> 
#> , , pob = B
#> 
#>          dest
#> orig              A        B        C         D     death outside
#>   A       17.058087  0.00000  0.00000  0.000000 3.9867110       0
#>   B        2.495671 28.68482 12.76128  5.780648 4.2307692       0
#>   C        0.000000  0.00000 18.72239  0.000000 5.1282051       0
#>   D        0.000000  0.00000  0.00000  9.417525 0.3636364       0
#>   birth    8.960573 15.94982 14.33692 10.752688 0.0000000       0
#>   outside  0.000000  0.00000  0.00000  0.000000 0.0000000       0
#> 
#> , , pob = C
#> 
#>          dest
#> orig              A        B          C        D     death outside
#>   A        9.318430  0.00000   0.000000  0.00000  1.993355       0
#>   B       10.706236 23.92586   4.876536  0.00000  3.076923       0
#>   C        0.000000  0.00000 114.549126  0.00000 28.717949       0
#>   D        7.192126  0.00000   3.275910 56.41139  2.363636       0
#>   birth    0.000000  0.00000  25.000000  0.00000  0.000000       0
#>   outside  0.000000  0.00000   0.000000  0.00000  0.000000       0
#> 
#> , , pob = D
#> 
#>          dest
#> orig              A        B         C        D    death outside
#>   A       17.092733  0.00000  0.000000   0.0000 3.986711       0
#>   B        2.132977 19.46418  1.049853   0.0000 1.923077       0
#>   C        0.000000  0.00000 15.008335   0.0000 4.102564       0
#>   D       10.296176  0.00000  5.067787 173.3691 7.272727       0
#>   birth    0.000000  0.00000  0.000000  60.0000 0.000000       0
#>   outside  0.000000  0.00000  0.000000   0.0000 0.000000       0
#> 
#> 
#> $stayer_assumption
#> [1] TRUE
#> 
#> $match_pob_tot_method
#> [1] "rescale"
#> 
#> $birth_non_negative
#> [1] TRUE
#> 
#> $death_method
#> [1] "proportion"
#> 
#> $od_flow
#>      dest
#> orig         A        B         C        D       Sum
#>   A    0.00000 8.258457  0.000000 0.000000  8.258457
#>   B   15.33488 0.000000 18.687672 5.780648 39.803204
#>   C    0.00000 1.051386  0.000000 0.000000  1.051386
#>   D   17.48830 0.000000  8.343697 0.000000 25.831999
#>   Sum 32.82319 9.309843 27.031369 5.780648 74.945047
#>