Summary indices of migration impact

Summary indices of migration impact

index_impact(
  m,
  p,
  pop_col = "pop",
  reg_col = "region",
  orig_col = "orig",
  dest_col = "dest",
  flow_col = "flow",
  long = TRUE
)

Arguments

m
A matrix or data frame of origin-destination flows. For matrix the first and second dimensions correspond to origin and destination respectively. For a data frame ensure the correct column names are passed to orig_col, dest_col and flow_col.

p
A data frame or named vector for the total population. When data frame, column of populations labelled using pop_col and region names labelled reg_col.

pop_col
Character string of the population column name

reg_col
Character string of the region column name. Must match dimension names or values in origin and destination columns of m.

orig_col
Character string of the origin column name (when m is a data frame rather than a matrix)

dest_col
Character string of the destination column name (when m is a data frame rather than a matrix)

flow_col
Character string of the flow column name (when m is a data frame rather than a matrix)

long
Logical to return a long data frame with index values all in one column

Source

Bell, M., Blake, M., Boyle, P., Duke-Williams, O., Rees, P. H., Stillwell, J., & Hugo, G. J. (2002). Cross-national comparison of internal migration: issues and measures. Journal of the Royal Statistical Society: Series A (Statistics in Society), 165(3), 435–464. https://doi.org/10.1111/1467-985X.00247

Shryock, H. S., & Siegel, J. S. (1976). The Methods and Materials of Demography. (E. G. Stockwell (ed.); Condensed). Academic Press.

United Nations Department of Economic and Social Affairs Population Division. (1983). Methods of measuring internal migration. United Nations Publication. https://www.un.org/en/development/desa/population/publications/manual/migration/measuring-migration.asp

Value

A tibble with 4 summary measures where

effectivness

Migration effectiveness index (MEI) from Shryock et al. (1975). Values range between 0 and 100. High values indicate migration is an efficient mechanism of population redistribution, generating a large net migration. Conversely, low values denote that migration is closely balanced, leading to comparatively little redistribution.

anmr

Aggregate net migration rate from Bell et. al. (2002). The population weighted version of mei.

perference

Index of preference, given in UN DESA (1983). From Bachi (1957) and Shryock et al. (1975) - measures size of migration compared to expected flows based on unifrom migration. Can go from 0 to infinity

velocity

Index of velocity, given in UN DESA (1983). From Bogue, Shryock, Jr. & Hoermann (1957) - measures size of migration compared to expected flows based on population size alone. Can go from 0 to infinity

Examples

# single year
index_impact(
  m = subset(korea_reg, year == 2020),
  p = subset(korea_pop, year == 2020),
  pop_col = "population"
)

#> # A tibble: 4 x 2
#>   measure        value
#>   <chr>          <dbl>
#> 1 effectivness   7.67 
#> 2 anmr           0.375
#> 3 preference   375.   
#> 4 velocity      18.3  

# multiple years
library(dplyr)
library(tidyr)
library(purrr) 
korea_reg %>%
  nest(m = c(orig, dest, flow)) %>%
  left_join(korea_pop) %>%
  nest(p = c(region, population)) %>%
  mutate(i = map2(.x = m, .y = p, 
                  .f = ~index_impact(m = .x, p = .y, pop_col = "population", long = FALSE))) %>%
  select(-m, -p) %>%
  unnest(i)

#> Joining, by = "year"

#> # A tibble: 9 x 5
#>    year effectivness  anmr preference velocity
#>   <int>        <dbl> <dbl>      <dbl>    <dbl>
#> 1  2012         6.07 0.300       409.     20.2
#> 2  2013         5.72 0.271       371.     17.6
#> 3  2014         5.36 0.262       434.     21.2
#> 4  2015         7.73 0.383       459.     22.7
#> 5  2016         8.47 0.402       401.     19.0
#> 6  2017         7.99 0.372       417.     19.4
#> 7  2018         9.29 0.435       398.     18.7
#> 8  2019         6.94 0.319       391.     18.0
#> 9  2020         7.67 0.375       375.     18.3