Summary indices of migration intensity

Summary indices of migration intensity

index_intensity(mig_total = NULL, pop_total = NULL, n = NULL, long = TRUE)

Arguments

mig_total
Numeric value for the total number of migrations.

pop_total
Numeric value for the total population.

n
Numeric value for the number of regions used in the definition of migration for mig_total.

long
Logical to return a long data frame with index values all in one column

Source

Bell, M., Blake, M., Boyle, P., Duke-Williams, O., Rees, P. H., Stillwell, J., & Hugo, G. J. (2002). Cross-national comparison of internal migration: issues and measures. Journal of the Royal Statistical Society: Series A (Statistics in Society), 165(3), 435–464. https://doi.org/10.1111/1467-985X.00247

Courgeau, D. (1973). Migrants et migrations. Population, 28(1), 95–129. https://doi.org/10.2307/1530972

Bernard, A., Rowe, F., Bell, M., Ueffing, P., Charles-Edwards, E., & Zhu, Y. (2017). Comparing internal migration across the countries of Latin America: A multidimensional approach. Plos One, 12(3), e0173895. https://doi.org/10.1371/journal.pone.0173895

Value

A tibble with 2 summary measures where

cmp

Crude migration probability from Bell et. al. (2002), sometimes known as crude migration intensity, e.g. Bernard (2017)

courgeau_k

Intensity measure of Courgeau (1973)

Examples

# single year
library(dplyr)
m <- korea_reg %>%
  filter(year == 2020,
         orig != dest) %>%
  pull(flow) %>%
  sum()
m

#> [1] 2534114

p <- korea_pop %>%
  filter(year == 2020) %>%
  pull(population) %>%
  sum()
p

#> [1] 51829023

index_intensity(mig_total = m, pop_total = p, n = n_distinct(korea_pop$region))

#> # A tibble: 2 x 2
#>   measure    value
#>   <chr>      <dbl>
#> 1 cmp        4.89 
#> 2 courgeau_k 0.863

# multiple years
mm <- korea_reg %>%
  group_by(year) %>%
  filter(orig != dest) %>%
  summarise(m = sum(flow))
mm

#> # A tibble: 9 x 2
#>    year       m
#>   <int>   <int>
#> 1  2012 2512740
#> 2  2013 2423429
#> 3  2014 2507796
#> 4  2015 2551424
#> 5  2016 2453342
#> 6  2017 2410930
#> 7  2018 2429184
#> 8  2019 2384948
#> 9  2020 2534114

pp <- korea_pop %>%
  group_by(year) %>%
  summarise(p = sum(population))
pp

#> # A tibble: 9 x 2
#>    year        p
#>   <int>    <dbl>
#> 1  2012 50948272
#> 2  2013 51141463
#> 3  2014 51327916
#> 4  2015 51529338
#> 5  2016 51696216
#> 6  2017 51778544
#> 7  2018 51826059
#> 8  2019 51849861
#> 9  2020 51829023

library(purrr)
library(tidyr)
mm %>%
  left_join(pp) %>%
  mutate(i = map2(.x = m, .y = p,
                  .f = ~index_intensity(mig_total = .x,
                                        pop_total = .y,
                                        n = n_distinct(korea_pop$region),
                                        long = FALSE))) %>%
  unnest(cols = i)

#> Joining, by = "year"

#> # A tibble: 9 x 5
#>    year       m        p   cmp courgeau_k
#>   <int>   <int>    <dbl> <dbl>      <dbl>
#> 1  2012 2512740 50948272  4.93      0.870
#> 2  2013 2423429 51141463  4.74      0.836
#> 3  2014 2507796 51327916  4.89      0.862
#> 4  2015 2551424 51529338  4.95      0.874
#> 5  2016 2453342 51696216  4.75      0.838
#> 6  2017 2410930 51778544  4.66      0.822
#> 7  2018 2429184 51826059  4.69      0.827
#> 8  2019 2384948 51849861  4.60      0.812
#> 9  2020 2534114 51829023  4.89      0.863