Chord diagram for directional origin-destination data

Adaption of circlize::chordDiagramFromDataFrame() with defaults set to allow for more effective visualisation of directional origin-destination data

mig_chord(
  x,
  lab = NULL,
  lab_bend1 = NULL,
  lab_bend2 = NULL,
  label_size = 1,
  label_nudge = 0,
  axis_size = 0.8,
  axis_breaks = NULL,
  ...,
  no_labels = FALSE,
  no_axis = FALSE,
  clear_circos_par = TRUE,
  zero_margin = TRUE,
  start.degree = 90,
  gap.degree = 4,
  track.margin = c(-0.1, 0.1),
  points.overflow.warning = FALSE
)

Arguments

x
Data frame with origin in first column, destination in second column and bilateral measure in third column

lab
Named vector of labels for plot. If NULL will use names from d

lab_bend1
Named vector of bending labels for plot. Note line breaks do not work with facing = "bending" in circlize.

lab_bend2
Named vector of second row of bending labels for plot.

label_size
Font size of label text.

label_nudge
Numeric value to nudge labels towards (negative number) or away (positive number) the sector axis.

axis_size
Font size on axis labels.

axis_breaks
Numeric value for how often to add axis label breaks. Default not activated, uses default from circlize::circos.axis()

...
Arguments for circlize::chordDiagramFromDataFrame().

no_labels
Logical to indicate if to include plot labels. Set to FALSE by default.

no_axis
Logical to indicate if to include plot axis. Set to FALSE by default.

clear_circos_par
Logical to run circlize::circos.clear(). Set to TRUE by default. Set to FALSE if you wish to add further to the plot.

zero_margin
Set margins of the plotting graphics device to zero. Set to TRUE by default.

start.degree
Argument for circlize::circos.par().

gap.degree
Argument for circlize::chordDiagramFromDataFrame().

track.margin
Argument for circlize::chordDiagramFromDataFrame().

points.overflow.warning
Argument for circlize::chordDiagramFromDataFrame().

Value

Chord diagram based on first three columns of x. The function tweaks the defaults of circlize::chordDiagramFromDataFrame() for easier plotting of directional origin-destination data. Users can override these defaults and pass additional tweaks using any of the circlize::chordDiagramFromDataFrame() arguments.

The layout of the plots are designed to specifically work on plotting images into PDF devices with widths and heights of 7 inches (the default dimension when using the pdf function). See the end of the examples for converting PDFs to images.

Fitting all the labels on the page is usually the most time consuming task. Use the different label options, including line breaks, label_nudge, track height in preAllocateTracks and font sizes in label_size and axis_size to find the best fit. If none of the label options produce desirable results, plot your own using circlize::circos.text having set no_labels = TRUE and clear_circos_par = FALSE.

Examples

# \dontrun{
library(tidyverse)
library(countrycode)
# download Abel and Cohen (2019) estimates
f <- read_csv("https://ndownloader.figshare.com/files/26239945")

#> Error in (function (con, what, n = 1L, size = NA_integer_, signed = TRUE,     endian = .Platform$endian) {    if (!endian %in% c("big", "little", "swap"))         stop("invalid 'endian' argument")    if (is.character(con)) {        con <- file(con, "rb")        on.exit(close(con))    }    swap <- endian != .Platform$endian    if (!is.character(what) || is.na(what) || length(what) !=         1L || !any(what == c("numeric", "double", "integer",         "int", "logical", "complex", "character", "raw")))         what <- typeof(what)    .Internal(readBin(con, what, n, size, signed, swap))})(structure(5L, class = c("curl", "connection"), conn_id = <pointer: 0x00000000000012c0>),     raw(0), 131071L): Send failure: Connection was reset

# use dictionary to get region to region flows
d <- f %>%
  mutate(
    orig = countrycode(
      sourcevar = orig, custom_dict = dict_ims,
      origin = "iso3c", destination = "region"),
    dest = countrycode(
      sourcevar = dest, custom_dict = dict_ims,
      origin = "iso3c", destination = "region")
  ) %>%
group_by(year0, orig, dest) %>%
summarise_all(sum) %>%
ungroup()

#> Error in mutate(., orig = countrycode(sourcevar = orig, custom_dict = dict_ims,     origin = "iso3c", destination = "region"), dest = countrycode(sourcevar = dest,     custom_dict = dict_ims, origin = "iso3c", destination = "region")): object 'f' not found

d

#> Error in eval(expr, envir, enclos): object 'd' not found

# 2015-2020 pseudo-Bayesian estimates for plotting
pb <- d %>%
    filter(year0 == 2015) %>%
    mutate(flow = da_pb_closed/1e6) %>%
    select(orig, dest, flow)

#> Error in filter(., year0 == 2015): object 'd' not found

    
# pdf(file = "chord.pdf")
mig_chord(x = pb)

#> Error in mig_chord(x = pb): object 'pb' not found

# dev.off()
# file.show("chord.pdf")

# pass arguments to circlize::chordDiagramFromDataFrame
# pdf(file = "chord.pdf")
mig_chord(x = pb, 
          # order of regions
          order = rev(unique(pb$orig)),
          # spacing for labels
          preAllocateTracks = list(track.height = 0.3),
          # colours
          grid.col = c("blue", "royalblue", "navyblue", "skyblue", "cadetblue", "darkblue")
          ) 

#> Error in unique(pb$orig): object 'pb' not found

# dev.off()
# file.show("chord.pdf")

# multiple line labels to fit on longer labels
r <- pb %>%
  sum_region() %>%
  mutate(lab = str_wrap_n(string = region, n = 2)) %>%
  separate(col = lab, into = c("lab1", "lab2"), sep = "\n", remove = FALSE, fill = "right")

#> Error in format_migration_tibble(m = m, orig_col = orig_col, dest_col = dest_col,     flow_col = flow_col): object 'pb' not found

r

#> Error in eval(expr, envir, enclos): object 'r' not found

# pdf(file = "chord.pdf")
mig_chord(x = pb,
          lab = r %>%
            select(region, lab) %>%
            deframe(),
          preAllocateTracks = list(track.height = 0.25),
          label_size = 0.8,
          axis_size = 0.7
          )

#> Error in mig_chord(x = pb, lab = r %>% select(region, lab) %>% deframe(),     preAllocateTracks = list(track.height = 0.25), label_size = 0.8,     axis_size = 0.7): object 'pb' not found

# dev.off()
# file.show("chord.pdf")

# bending labels
# pdf(file = "chord.pdf")
mig_chord(x = pb,
          lab_bend1 = r %>%
            select(region, lab1) %>%
            deframe(),
          lab_bend2 = r %>%
            select(region, lab2) %>%
            deframe()
          )

#> Error in mig_chord(x = pb, lab_bend1 = r %>% select(region, lab1) %>%     deframe(), lab_bend2 = r %>% select(region, lab2) %>% deframe()): object 'pb' not found

# dev.off()
# file.show("chord.pdf")

# convert pdf to image file
# library(magick)
# p <- image_read_pdf("chord.pdf")
# image_write(image = p, path = "chord.png")
# file.show("chord.png")
# }