Estimate net migration from survival ratios applied to lifetime migration data

Using survival ratios to estimate net migration from lifetime migration data

Usage

net_sr(
  .data,
  pop0_col = "pop0",
  pop1_col = "pop1",
  survival_ratio_col = "sr",
  net_children = FALSE,
  maternal_exposure = c(0.25, 0.75),
  maternal_age_id = 4:9,
  maternal_col = pop1_col
)

Arguments

.data

A data frame with two rows with the total number of lifetime in- and out-migrants in separate columns. The first row contains totals at the first time point and second row at the second time point.

pop0_col

Character string name of column containing name of initial populations. Default "pop0".

pop1_col

Character string name of column containing name of end populations. Default "pop1".

survival_ratio_col

Character string name of column containing survivor ratios. Default "sr".

net_children

Logical to indicate if to estimate net migration when no survival ratio exists. Default `FALSE`.

maternal_exposure

Vector for maternal exposures to interval to be used to estimate net migration for each of the unknown children age groups. Length should correspond to the number of children age groups where net migration estimates are required.

maternal_age_id

Row numbers to indicate which rows correspond to maternal age groups at the end of the period.

maternal_col

Name of maternal population column, required for the estimation of net migration of children.

Value

Data frame with estimates of net migration

References

Bogue, D. J., Hinze, K., & White, M. (1982). Techniques of Estimating Net Migration. Community and Family Study Center. University of Chicago.

Examples

# results to match un manual 1984 (table 24)
net_sr(bombay_1951, pop0_col = "pop_1941", pop1_col = "pop_1951")
#> # A tibble: 13 × 10
#>    age_1…¹ age_1…² pop_1…³ pop_1…⁴    sr net_f…⁵ net_r…⁶ net_a…⁷ pop1_…⁸ pop0_…⁹
#>    <fct>   <fct>     <dbl>   <dbl> <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>
#>  1 0-4     10-14     77135  132870 0.909  62777.  69085.  65931.  70093. 146220.
#>  2 5-9     15-19     85434  170227 0.957  88441.  92386.  90413.  81786. 177820.
#>  3 10-14   20-24     79185  263971 0.947 188975. 199530. 194252.  74996. 278715.
#>  4 15-19   25-29     82603  253964 0.931 177077. 190242. 183659.  76887. 272845.
#>  5 20-24   30-34    126247  195373 0.922  78935.  85585.  82260. 116438. 211832.
#>  6 25-29   35-39    155344  151259 0.916   8948.   9768.   9358. 142311. 165112.
#>  7 30-34   40-44    138843  118383 0.905  -7228.  -7990.  -7609. 125611. 130853.
#>  8 35-39   45-49    109356   76421 0.885 -20359. -23005. -21682.  96780.  86351.
#>  9 40-44   50-54     81626   65897 0.855  -3877.  -4535.  -4206.  69774.  77091.
#> 10 45-49   55-59     47062   32265 0.812  -5959.  -7337.  -6648.  38224.  39725.
#> 11 50-54   60-64     36908   22248 0.754  -5562.  -7382.  -6472.  27810.  29526.
#> 12 55-59   65-69     15134    9655 0.673   -524.   -779.   -652.  10179.  14355.
#> 13 60+     70+       25094   10100 0.387    399.   1031.    715.   9701.  26125.
#> # … with abbreviated variable names ¹​age_1941, ²​age_1951, ³​pop_1941, ⁴​pop_1951,
#> #   ⁵​net_forward, ⁶​net_reverse, ⁷​net_average, ⁸​pop1_forward, ⁹​pop0_reverse
  
# results to match Bogue, Hinze and White (1982)
library(dplyr)
alabama_1970 %>%
  filter(race == "white", sex == "male") %>%
  select(-race, -sex) %>%
  group_by(age_1970) %>%
  net_sr(pop0_col = "pop_1960", pop1_col = "pop_1970", 
         survival_ratio_col = "us_census_sr")
#> # A tibble: 17 × 9
#>    age_1970 pop_1960 pop_1970 us_censu…¹ net_f…² net_r…³ net_a…⁴ pop1_…⁵ pop0_…⁶
#>    <fct>       <dbl>    <dbl>      <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>
#>  1 0-4        110365   105503      0.955  5.70e1  5.96e1  5.83e1 105446. 110425.
#>  2 5-9        126116   120884      0.947  1.48e3  1.56e3  1.52e3 119403. 127680.
#>  3 10-14      126346   128656      1.00   2.06e3  2.06e3  2.06e3 126593. 128405.
#>  4 15-19      119571   118485      1.01  -1.70e3 -1.70e3 -1.70e3 120189. 117875.
#>  5 20-24      117525   102296      0.989 -1.39e4 -1.41e4 -1.40e4 116225. 103440.
#>  6 25-29       94148    85778      0.973 -5.85e3 -6.01e3 -5.93e3  91624.  88141.
#>  7 30-34       72012    74046      0.985  3.14e3  3.19e3  3.16e3  70908.  75199.
#>  8 35-39       70867    71636      0.980  2.18e3  2.22e3  2.20e3  69460.  73087.
#>  9 40-44       75168    75179      0.972  2.13e3  2.19e3  2.16e3  73052.  77357.
#> 10 45-49       76872    73602      0.948  7.08e2  7.46e2  7.27e2  72894.  77618.
#> 11 50-54       72068    67679      0.933  4.66e2  4.99e2  4.82e2  67213.  72567.
#> 12 55-59       69002    60807      0.885 -2.73e2 -3.08e2 -2.91e2  61080.  68694.
#> 13 60-64       60654    52073      0.845  7.91e2  9.36e2  8.63e2  51282.  61590.
#> 14 65-69       49127    38786      0.750  1.92e3  2.55e3  2.24e3  36869.  51682.
#> 15 70-74       37344    26632      0.675  1.44e3  2.13e3  1.78e3  25196.  39473.
#> 16 75-79       32063    17418      0.537  2.14e2  3.99e2  3.06e2  17204.  32462.
#> 17 80+         51294    16029      0.296  8.60e2  2.91e3  1.88e3  15169.  54202.
#> # … with abbreviated variable names ¹​us_census_sr, ²​net_forward, ³​net_reverse,
#> #   ⁴​net_average, ⁵​pop1_forward, ⁶​pop0_reverse
         
# results to match UN manual 1992 (table 6)
net_sr(manila_1970, pop0_col = "pop_1960", pop1_col = "pop_1970", 
       survival_ratio_col = "phl_census_sr")
#> # A tibble: 16 × 9
#>    age_1970 pop_1960 pop_1970 phl_cens…¹ net_f…² net_r…³ net_a…⁴ pop1_…⁵ pop0_…⁶
#>    <fct>       <dbl>    <dbl>      <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>
#>  1 0-4            NA    85870     NA          0       0       0      NA      NA 
#>  2 5-9            NA    83054     NA          0       0       0      NA      NA 
#>  3 10-14       80275    79489      1.12  -10196.  -9126.  -9661.  89685.  71149.
#>  4 15-19       70875   101410      0.992  31134.  31400.  31267.  70276. 102275.
#>  5 20-24       63250    90410      0.973  28877.  29683.  29280.  61533.  92933.
#>  6 25-29       85618    56055      0.889 -20082. -22582. -21332.  76137.  63036.
#>  7 30-34       75793    44648      0.841 -19117. -22723. -20920.  63765.  53070.
#>  8 35-39       60037    36963      0.957 -20497. -21416. -20957.  57460.  38621.
#>  9 40-44       34813    28873      0.951  -4244.  -4462.  -4353.  33117.  30351.
#> 10 45-49       31927    23678      0.904  -5189.  -5739.  -5464.  28867.  26188.
#> 11 50-54       24297    19063      0.930  -3521.  -3788.  -3655.  22584.  20509.
#> 12 55-59       20207    14484      0.797  -1613.  -2025.  -1819.  16097.  18182.
#> 13 60-64       13714    10205      0.877  -1822.  -2078.  -1950.  12027.  11636.
#> 14 65-69        9366     6405      0.835  -1417.  -1697.  -1557.   7822.   7669.
#> 15 70-74        7921     3746      0.712  -1890.  -2657.  -2274.   5636.   5264.
#> 16 75+         11114     4779      0.562  -1472.  -2617.  -2045.   6251.   8497.
#> # … with abbreviated variable names ¹​phl_census_sr, ²​net_forward, ³​net_reverse,
#> #   ⁴​net_average, ⁵​pop1_forward, ⁶​pop0_reverse
       
# with children net migration estimate
net_sr(manila_1970, pop0_col = "pop_1960", pop1_col = "pop_1970", 
       survival_ratio_col = "phl_census_sr", net_children = TRUE)
#> # A tibble: 16 × 9
#>    age_1970 pop_1960 pop_1970 phl_cens…¹ net_f…² net_r…³ net_a…⁴ pop1_…⁵ pop0_…⁶
#>    <fct>       <dbl>    <dbl>      <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>
#>  1 0-4            NA    85870     NA       -235.   -605.   -420.     NA      NA 
#>  2 5-9            NA    83054     NA      -8935. -10486.  -9710.     NA      NA 
#>  3 10-14       80275    79489      1.12  -10196.  -9126.  -9661.  89685.  71149.
#>  4 15-19       70875   101410      0.992  31134.  31400.  31267.  70276. 102275.
#>  5 20-24       63250    90410      0.973  28877.  29683.  29280.  61533.  92933.
#>  6 25-29       85618    56055      0.889 -20082. -22582. -21332.  76137.  63036.
#>  7 30-34       75793    44648      0.841 -19117. -22723. -20920.  63765.  53070.
#>  8 35-39       60037    36963      0.957 -20497. -21416. -20957.  57460.  38621.
#>  9 40-44       34813    28873      0.951  -4244.  -4462.  -4353.  33117.  30351.
#> 10 45-49       31927    23678      0.904  -5189.  -5739.  -5464.  28867.  26188.
#> 11 50-54       24297    19063      0.930  -3521.  -3788.  -3655.  22584.  20509.
#> 12 55-59       20207    14484      0.797  -1613.  -2025.  -1819.  16097.  18182.
#> 13 60-64       13714    10205      0.877  -1822.  -2078.  -1950.  12027.  11636.
#> 14 65-69        9366     6405      0.835  -1417.  -1697.  -1557.   7822.   7669.
#> 15 70-74        7921     3746      0.712  -1890.  -2657.  -2274.   5636.   5264.
#> 16 75+         11114     4779      0.562  -1472.  -2617.  -2045.   6251.   8497.
#> # … with abbreviated variable names ¹​phl_census_sr, ²​net_forward, ³​net_reverse,
#> #   ⁴​net_average, ⁵​pop1_forward, ⁶​pop0_reverse