Expand flow data to include aggregate sums for origin and destination meta-regions

Expand matrix of data frame of migration data to include aggregate sums for corresponding origin and destination meta regions.

sum_expand(
  m,
  return_matrix = TRUE,
  guess_order = TRUE,
  area_first = TRUE,
  orig_col = "orig",
  dest_col = "dest",
  flow_col = "flow",
  orig_area_col = "orig_area",
  dest_area_col = "dest_area",
  orig_area = NULL,
  dest_area = NULL
)

Arguments

m
A matrix or data frame of origin-destination flows. For matrix the first and second dimensions correspond to origin and destination respectively. For a data frame ensure the correct column names are passed to orig_col, dest_col and flow_col.

return_matrix
Logical to return a matrix. Default TRUE.

guess_order
Logical to return a matrix or data frame ordered by origin and destination with area names at the end of each block. Default TRUE. If FALSE returns matrix or data frame based on alphabetical order of origin and destinations.

area_first
Order area sums to be placed before the origin and destination values. Default TRUE

orig_col
Character string of the origin column name (when m is a data frame rather than a matrix)

dest_col
Character string of the destination column name (when m is a data frame rather than a matrix)

flow_col
Character string of the flow column name (when m is a data frame rather than a matrix)

orig_area_col
Character string of the origin area column name (when m is a data frame rather than a matrix)

dest_area_col
Character string of the destination area column name (when m is a data frame rather than a matrix)

orig_area
Vector of labels for the origin areas of each row of m.

dest_area
Vector of labels for the destination areas of each row of m.

Value

A tibble or matrix with additional row and columns (for matrices) for aggregate sums for origin and destination meta-regions

Examples

##
## from matrix
##
m <- block_matrix(x = 1:16, b = c(2,3,4,2))
m

#>    A1 A2 B1 B2 B3 C1 C2 C3 C4 D1 D2
#> A1  1  1  5  5  5  9  9  9  9 13 13
#> A2  1  1  5  5  5  9  9  9  9 13 13
#> B1  2  2  6  6  6 10 10 10 10 14 14
#> B2  2  2  6  6  6 10 10 10 10 14 14
#> B3  2  2  6  6  6 10 10 10 10 14 14
#> C1  3  3  7  7  7 11 11 11 11 15 15
#> C2  3  3  7  7  7 11 11 11 11 15 15
#> C3  3  3  7  7  7 11 11 11 11 15 15
#> C4  3  3  7  7  7 11 11 11 11 15 15
#> D1  4  4  8  8  8 12 12 12 12 16 16
#> D2  4  4  8  8  8 12 12 12 12 16 16

# requires a vector of origin and destination areas
a <- rep(LETTERS[1:4], times = c(2,3,4,2))
a

#>  [1] "A" "A" "B" "B" "B" "C" "C" "C" "C" "D" "D"

sum_expand(m = m, orig_area = a, dest_area = a)

#>     dest
#> orig  A A1 A2  B B1 B2 B3   C C1 C2 C3 C4   D D1 D2
#>   A   4  2  2 30 10 10 10  72 18 18 18 18  52 26 26
#>   A1  2  1  1 15  5  5  5  36  9  9  9  9  26 13 13
#>   A2  2  1  1 15  5  5  5  36  9  9  9  9  26 13 13
#>   B  12  6  6 54 18 18 18 120 30 30 30 30  84 42 42
#>   B1  4  2  2 18  6  6  6  40 10 10 10 10  28 14 14
#>   B2  4  2  2 18  6  6  6  40 10 10 10 10  28 14 14
#>   B3  4  2  2 18  6  6  6  40 10 10 10 10  28 14 14
#>   C  24 12 12 84 28 28 28 176 44 44 44 44 120 60 60
#>   C1  6  3  3 21  7  7  7  44 11 11 11 11  30 15 15
#>   C2  6  3  3 21  7  7  7  44 11 11 11 11  30 15 15
#>   C3  6  3  3 21  7  7  7  44 11 11 11 11  30 15 15
#>   C4  6  3  3 21  7  7  7  44 11 11 11 11  30 15 15
#>   D  16  8  8 48 16 16 16  96 24 24 24 24  64 32 32
#>   D1  8  4  4 24  8  8  8  48 12 12 12 12  32 16 16
#>   D2  8  4  4 24  8  8  8  48 12 12 12 12  32 16 16

# place area sums after regions
sum_expand(m = m, orig_area = a, dest_area = a, area_first = FALSE)

#>     dest
#> orig A1 A2  A B1 B2 B3  B C1 C2 C3 C4   C D1 D2   D
#>   A1  1  1  2  5  5  5 15  9  9  9  9  36 13 13  26
#>   A2  1  1  2  5  5  5 15  9  9  9  9  36 13 13  26
#>   A   2  2  4 10 10 10 30 18 18 18 18  72 26 26  52
#>   B1  2  2  4  6  6  6 18 10 10 10 10  40 14 14  28
#>   B2  2  2  4  6  6  6 18 10 10 10 10  40 14 14  28
#>   B3  2  2  4  6  6  6 18 10 10 10 10  40 14 14  28
#>   B   6  6 12 18 18 18 54 30 30 30 30 120 42 42  84
#>   C1  3  3  6  7  7  7 21 11 11 11 11  44 15 15  30
#>   C2  3  3  6  7  7  7 21 11 11 11 11  44 15 15  30
#>   C3  3  3  6  7  7  7 21 11 11 11 11  44 15 15  30
#>   C4  3  3  6  7  7  7 21 11 11 11 11  44 15 15  30
#>   C  12 12 24 28 28 28 84 44 44 44 44 176 60 60 120
#>   D1  4  4  8  8  8  8 24 12 12 12 12  48 16 16  32
#>   D2  4  4  8  8  8  8 24 12 12 12 12  48 16 16  32
#>   D   8  8 16 16 16 16 48 24 24 24 24  96 32 32  64

##
## from large data frame
##
# \dontrun{
library(tidyverse)

#> -- Attaching packages --------------------------------------- tidyverse 1.3.1 --

#> v ggplot2 3.3.5     v stringr 1.4.0
#> v readr   1.4.0     v forcats 0.5.1

#> -- Conflicts ------------------------------------------ tidyverse_conflicts() --
#> x dplyr::filter() masks stats::filter()
#> x dplyr::lag()    masks stats::lag()

library(countrycode)

# download Abel and Cohen (2019) estimates
f <- read_csv("https://ndownloader.figshare.com/files/26239945")

#> 
#> -- Column specification --------------------------------------------------------
#> cols(
#>   year0 = col_double(),
#>   orig = col_character(),
#>   dest = col_character(),
#>   sd_drop_neg = col_double(),
#>   sd_rev_neg = col_double(),
#>   mig_rate = col_double(),
#>   da_min_open = col_double(),
#>   da_min_closed = col_double(),
#>   da_pb_closed = col_double()
#> )

cm <- c("CHI" = "Europe",
        "SCG" = "Europe", 
        "SUD" = "Africa")

# 1990-1995 flow estimates
f %>%
  filter(year0 == 1990) %>%
  mutate(
    orig_area = countrycode(sourcevar = orig, custom_match = cm,
                            origin = "iso3c", destination = "un.region.name"),
    dest_area = countrycode(sourcevar = dest, custom_match = cm,
                            origin = "iso3c", destination = "un.region.name")
  ) %>%
  sum_expand(flow_col = "da_pb_closed", return_matrix = FALSE)

#> # A tibble: 40,804 x 3
#>    orig   dest      flow
#>    <chr>  <chr>    <dbl>
#>  1 Africa Africa 9733640
#>  2 Africa BDI      60436
#>  3 Africa COM       6434
#>  4 Africa DJI      10274
#>  5 Africa ERI      10938
#>  6 Africa ETH    1421749
#>  7 Africa KEN     238693
#>  8 Africa MDG       2743
#>  9 Africa MWI     105518
#> 10 Africa MUS        788
#> # ... with 40,794 more rows

# by group (period)
f %>%
  mutate(
    orig_area = countrycode(sourcevar = orig, custom_match = cm,
                            origin = "iso3c", destination = "un.region.name"),
    dest_area = countrycode(sourcevar = dest, custom_match = cm,
                            origin = "iso3c", destination = "un.region.name")) %>%
  group_by(year0) %>%
  sum_expand(flow_col = "da_pb_closed", return_matrix = FALSE)

#> # A tibble: 247,266 x 4
#> # Groups:   year0 [6]
#>    year0 orig   dest      flow
#>    <dbl> <chr>  <chr>    <dbl>
#>  1  1990 Africa Africa 9733640
#>  2  1995 Africa Africa 7366877
#>  3  2000 Africa Africa 6642723
#>  4  2005 Africa Africa 6654457
#>  5  2010 Africa Africa 8776877
#>  6  2015 Africa Africa 8455447
#>  7  1990 Africa BDI      60436
#>  8  1995 Africa BDI      64031
#>  9  2000 Africa BDI     117710
#> 10  2005 Africa BDI     158228
#> # ... with 247,256 more rows

# }