Summary of regional in-, out-, turnover and net-migration totals from an origin-destination migration flow matrix or data frame.

Summary of regional in-, out-, turnover and net-migration totals from an origin-destination migration flow matrix or data frame.

sum_region(
  m,
  drop_diagonal = TRUE,
  orig_col = "orig",
  dest_col = "dest",
  flow_col = "flow",
  international = FALSE,
  include_net = TRUE
)

sum_country(
  m,
  drop_diagonal = TRUE,
  orig_col = "orig",
  dest_col = "dest",
  flow_col = "flow",
  include_net = TRUE,
  international = TRUE
)

Arguments

m

A matrix or data frame of origin-destination flows. For matrix the first and second dimensions correspond to origin and destination respectively. For a data frame ensure the correct column names are passed to orig_col, dest_col and flow_col.

drop_diagonal

Logical to indicate dropping of diagonal terms, where the origin and destination are the same, in the calculation of totals. Default TRUE.

orig_col

Character string of the origin column name (when m is a data frame rather than a matrix)

dest_col

Character string of the destination column name (when m is a data frame rather than a matrix)

flow_col

Character string of the flow column name (when m is a data frame rather than a matrix)

international

Logical to indicate if flows are international.

include_net

Logical to indicate inclusion of a net migration total column for each region, in addition to the total in- and out-flows. Default TRUE.

Value

A tibble with total in-, out- and turnover of flows for each region.

Examples

# matrix
r <- LETTERS[1:4]
m <- matrix(data = c(0, 100, 30, 70, 50, 0, 45, 5, 60, 35, 0, 40, 20, 25, 20, 0),
            nrow = 4, ncol = 4, dimnames = list(orig = r, dest = r), byrow = TRUE)
m
#>     dest
#> orig  A   B  C  D
#>    A  0 100 30 70
#>    B 50   0 45  5
#>    C 60  35  0 40
#>    D 20  25 20  0
sum_region(m)
#> # A tibble: 4 × 5
#>   region out_mig in_mig  turn   net
#>   <chr>    <dbl>  <dbl> <dbl> <dbl>
#> 1 A          200    130   330   -70
#> 2 B          100    160   260    60
#> 3 C          135     95   230   -40
#> 4 D           65    115   180    50
  
# \dontrun{
# data frame (tidy) format
library(tidyverse)

# download Abel and Cohen (2019) estimates
f <- read_csv("https://ndownloader.figshare.com/files/26239945", show_col_types = FALSE)
f
#> # A tibble: 235,236 × 9
#>    year0 orig  dest  sd_drop_neg sd_rev_neg mig_rate da_min_open da_mi…¹ da_pb…²
#>    <dbl> <chr> <chr>       <dbl>      <dbl>    <dbl>       <dbl>   <dbl>   <dbl>
#>  1  1990 BDI   BDI             0          0        0           0       0       0
#>  2  1990 COM   BDI             0          0        0           0       0       0
#>  3  1990 DJI   BDI             0          0        0           0       0       0
#>  4  1990 ERI   BDI             0          0        0           0       0      90
#>  5  1990 ETH   BDI             0          0        0           0       0       2
#>  6  1990 KEN   BDI            30         30       69          45      29      87
#>  7  1990 MDG   BDI             0          0        0           0       0       0
#>  8  1990 MWI   BDI             0          0        0           0       0     125
#>  9  1990 MUS   BDI             0          0        0           0       1       1
#> 10  1990 MYT   BDI             0          0        0           0       0       0
#> # … with 235,226 more rows, and abbreviated variable names ¹​da_min_closed,
#> #   ²​da_pb_closed

# single period
f %>% 
  filter(year0 == 1990) %>%
  sum_country(flow_col = "da_pb_closed")
#> # A tibble: 197 × 5
#>    country    emi     imm    turn     net
#>    <chr>    <dbl>   <dbl>   <dbl>   <dbl>
#>  1 ABW       1662   15874   17536   14212
#>  2 AFG     345255 3421712 3766967 3076457
#>  3 AGO      82775  225637  308412  142862
#>  4 ALB     464693   21479  486172 -443214
#>  5 ARE     272648  640784  913432  368136
#>  6 ARG     444239  339393  783632 -104846
#>  7 ARM     648202  151937  800139 -496265
#>  8 ATG       6153    8387   14540    2234
#>  9 AUS     691618 1042781 1734399  351163
#> 10 AUT     154853  382724  537577  227871
#> # … with 187 more rows

# all periods using group_by
f %>% 
  group_by(year0) %>%
  sum_country(flow_col = "da_pb_closed")
#> # A tibble: 1,188 × 6
#> # Groups:   year0 [6]
#>    year0 country    emi     imm    turn     net
#>    <dbl> <chr>    <dbl>   <dbl>   <dbl>   <dbl>
#>  1  1990 ABW       1662   15874   17536   14212
#>  2  1990 AFG     345255 3421712 3766967 3076457
#>  3  1990 AGO      82775  225637  308412  142862
#>  4  1990 ALB     464693   21479  486172 -443214
#>  5  1990 ARE     272648  640784  913432  368136
#>  6  1990 ARG     444239  339393  783632 -104846
#>  7  1990 ARM     648202  151937  800139 -496265
#>  8  1990 ATG       6153    8387   14540    2234
#>  9  1990 AUS     691618 1042781 1734399  351163
#> 10  1990 AUT     154853  382724  537577  227871
#> # … with 1,178 more rows
# }