Calculate turnover and total in-, out- and net-migration totals from an origin-destination migration flow matrix or data frame.

Calculate turnover and total in-, out- and net-migration totals from an origin-destination migration flow matrix or data frame.

sum_turnover(
  m,
  drop_diagonal = TRUE,
  include_net = TRUE,
  orig_col = "orig",
  dest_col = "dest",
  flow_col = "flow",
  type = "internal",
  international = FALSE,
  name_tot = FALSE
)

Arguments

m
A matrix or data frame of origin-destination flows. For matrix the first and second dimensions correspond to origin and destination respectively. For a data frame ensure the correct column names are passed to orig_col, dest_col and flow_col.

drop_diagonal
Logical to indicate dropping of diagonal terms, where the origin and destination are the same, in the calculation of totals. Default TRUE.

include_net
Logical to indicate inclusion of a net migration total column for each region, in addition to the total in- and out-flows. Default TRUE.

orig_col
Character string of the origin column name (when m is a data frame rather than a matrix)

dest_col
Character string of the destination column name (when m is a data frame rather than a matrix)

flow_col
Character string of the flow column name (when m is a data frame rather than a matrix)

type
Character string to indicate if flows are internal or international to indicate if to use region, tot_in_mig, tot_out_mig or country, tot_imm and tot_emi in output.

international
Logical to indicate if flows are international.

name_tot
Logical to prefix column names with `"tot_"` to match older versions. Default `FALSE`.

Value

A tibble with total in-, out- and turnover of flows for each region.

Examples

# matrix
r <- LETTERS[1:4]
m <- matrix(data = c(0, 100, 30, 70, 50, 0, 45, 5, 60, 35, 0, 40, 20, 25, 20, 0),
            nrow = 4, ncol = 4, dimnames = list(orig = r, dest = r), byrow = TRUE)
m

#>     dest
#> orig  A   B  C  D
#>    A  0 100 30 70
#>    B 50   0 45  5
#>    C 60  35  0 40
#>    D 20  25 20  0

sum_turnover(m)

#> # A tibble: 4 x 5
#>   region out_mig in_mig  turn   net
#>   <chr>    <dbl>  <dbl> <dbl> <dbl>
#> 1 A          200    130   330   -70
#> 2 B          100    160   260    60
#> 3 C          135     95   230   -40
#> 4 D           65    115   180    50

  
# different labels
sum_turnover(m, international = TRUE)

#> # A tibble: 4 x 5
#>   country   emi   imm  turn   net
#>   <chr>   <dbl> <dbl> <dbl> <dbl>
#> 1 A         200   130   330   -70
#> 2 B         100   160   260    60
#> 3 C         135    95   230   -40
#> 4 D          65   115   180    50

# \dontrun{
# data frame (tidy) format
library(tidyverse)

# download Abel and Cohen (2019) estimates
f <- read_csv("https://ndownloader.figshare.com/files/26239945")

#> 
#> -- Column specification --------------------------------------------------------
#> cols(
#>   year0 = col_double(),
#>   orig = col_character(),
#>   dest = col_character(),
#>   sd_drop_neg = col_double(),
#>   sd_rev_neg = col_double(),
#>   mig_rate = col_double(),
#>   da_min_open = col_double(),
#>   da_min_closed = col_double(),
#>   da_pb_closed = col_double()
#> )

# turnover for single period
f %>% 
  filter(year0 == 1990) %>%
  sum_turnover(flow_col = "da_pb_closed", type = "international")

#> # A tibble: 197 x 5
#>    country    emi     imm    turn     net
#>    <chr>    <dbl>   <dbl>   <dbl>   <dbl>
#>  1 ABW       1662   15874   17536   14212
#>  2 AFG     345255 3421712 3766967 3076457
#>  3 AGO      82775  225637  308412  142862
#>  4 ALB     464693   21479  486172 -443214
#>  5 ARE     272648  640784  913432  368136
#>  6 ARG     444239  339393  783632 -104846
#>  7 ARM     648202  151937  800139 -496265
#>  8 ATG       6153    8387   14540    2234
#>  9 AUS     691618 1042781 1734399  351163
#> 10 AUT     154853  382724  537577  227871
#> # ... with 187 more rows

# turnover for all periods using group_by
f %>% 
  group_by(year0) %>%
  sum_turnover(flow_col = "da_pb_closed", type = "international")

#> # A tibble: 1,188 x 6
#> # Groups:   year0 [6]
#>    year0 country    emi     imm    turn     net
#>    <dbl> <chr>    <dbl>   <dbl>   <dbl>   <dbl>
#>  1  1990 ABW       1662   15874   17536   14212
#>  2  1990 AFG     345255 3421712 3766967 3076457
#>  3  1990 AGO      82775  225637  308412  142862
#>  4  1990 ALB     464693   21479  486172 -443214
#>  5  1990 ARE     272648  640784  913432  368136
#>  6  1990 ARG     444239  339393  783632 -104846
#>  7  1990 ARM     648202  151937  800139 -496265
#>  8  1990 ATG       6153    8387   14540    2234
#>  9  1990 AUS     691618 1042781 1734399  351163
#> 10  1990 AUT     154853  382724  537577  227871
#> # ... with 1,178 more rows

# }